+++ public/static/lang/en.json
@@ -558,11 +558,11 @@
   "DateRange.LastWeek": "Last Week",
   "DateRange.LastMonth": "Last Month",
   "DateRange.LastYear": "Last Year",
-  "ThirdPartyGraph.Label.gtr": "Google (Search)",
-  "ThirdPartyGraph.Label.merit-nt": "Telescope",
-  "ThirdPartyGraph.Label.bgp": "BGP",
-  "ThirdPartyGraph.Label.ping-slash24": "Active Probing",
-  "ThirdPartyGraph.Label.cloudflare": "Cloudflare",
+  "ThirdPartyChart.Label.gtr": "Google (Search)",
+  "ThirdPartyChart.Label.merit-nt": "Telescope",
+  "ThirdPartyChart.Label.bgp": "BGP",
+  "ThirdPartyChart.Label.ping-slash24": "Active Probing",
+  "ThirdPartyChart.Label.cloudflare": "Cloudflare",
   "Highlights.Political.CubaReferendum2019.Title": "2019 Constitutional Referendum",
   "Highlights.Political.CubaReferendum2019.Text": "Blocking of independent media",
   "Highlights.Political.VenezuelaCrisis2019.Title": "2019 Presidential Crisis",
