+++ public/static/lang/translations.js
@@ -1,1 +1,1 @@
-window.OONITranslations = {"en":{"General.OoniExplorer":"OONI Explorer","General.OK":"OK","General.Error":"Error","General.Anomaly":"Anomaly","General.Accessible":"Accessible","General.Failed":"Failed","General.Loading":"Loading…","General.NoData":"No data","General.Apply":"Apply","General.Reset":"Reset","General.Submit":"Submit","General.Close":"Close","General.Cancel":"Cancel","General.Login":"Login","General.Logout":"Logout","General.Edit":"Edit","Login.EnterEmail":"Add your email address and click the link sent to your email to log in.\n\nWe do not store email addresses.","Login.Submitted":"Your login request has been submitted. Please check your email for a link to activate and log in to your account.","Login.LoggingIn":"Logging in...","Login.Failure":"Try logging in again","Login.Success":"Successfully logged in. Redirecting back to the measurement...","SocialButtons.CTA":"Share on <facebook-link>Facebook</facebook-link> or <twitter-link>Twitter</twitter-link>","SocialButtons.Text":"Data from OONI Explorer","Tests.WebConnectivity.Name":"Web Connectivity Test","Tests.Telegram.Name":"Telegram Test","Tests.Facebook.Name":"Facebook Messenger Test","Tests.WhatsApp.Name":"WhatsApp Test","Tests.Signal.Name":"Signal Test","Tests.HTTPInvalidReqLine.Name":"HTTP Invalid Request Line Test","Tests.HTTPHeaderManipulation.Name":"HTTP Header Field Manipulation Test","Tests.NDT.Name":"NDT Speed Test","Tests.Dash.Name":"DASH Video Streaming Test","Tests.TorVanilla.Name":"Tor (Vanilla) Test","Tests.BridgeReachability.Name":"Tor Bridge Reachability Test","Tests.TCPConnect.Name":"TCP Connect Test","Tests.DNSConsistency.Name":"DNS Consistency Test","Tests.HTTPRequests.Name":"HTTP Requests Test","Tests.Psiphon.Name":"Psiphon Test","Tests.Tor.Name":"Tor Test","Tests.RiseupVPN.Name":"Riseup VPN Test","Tests.TorSnowflake.Name":"Tor Snowflake Test","Tests.DNSCheck.Name":"DNS Check","Tests.StunReachability.Name":"STUN Reachability","Tests.URLGetter.Name":"URL Getter","Tests.Groups.Webistes.Name":"Websites","Tests.Groups.Instant Messagging.Name":"Instant Messaging","Tests.Groups.Middlebox.Name":"Middleboxes","Tests.Groups.Performance.Name":"Performance","Tests.Groups.Circumvention.Name":"Circumvention","Tests.Groups.Experimental.Name":"Experimental","Tests.Groups.Legacy.Name":"Legacy","Measurement.MetaDescription":"OONI data suggests {description} on {formattedDate}, find more open data on internet censorship on OONI Explorer.","Measurement.NotFound":"Measurement not found","Measurement.Hero.Status.Confirmed":"Confirmed Blocked","Measurement.Hero.Status.Down":"Website Down","Measurement.Hero.Status.Anomaly.DNS":"DNS","Measurement.Hero.Status.Anomaly.HTTP":"HTTP","Measurement.Hero.Status.Anomaly.TCP":"TCP/IP","Measurement.CommonSummary.Label.ASN":"Network","Measurement.CommonSummary.Label.Country":"Country","Measurement.CommonSummary.Label.DateTime":"Date & Time","Measurement.DetailsHeader.Runtime":"Runtime","Measurement.Status.Hint.Websites.Censorship":"","Measurement.Status.Hint.Websites.DNS":"DNS tampering","Measurement.Status.Hint.Websites.Error":"Error in detection","Measurement.Status.Hint.Websites.HTTPdiff":"HTTP blocking (a blockpage might be served)","Measurement.Status.Hint.Websites.HTTPfail":"HTTP blocking (HTTP requests failed)","Measurement.Status.Hint.Websites.NoCensorship":"No blocking detected","Measurement.Status.Hint.Websites.TCPBlock":"TCP/IP blocking","Measurement.Status.Hint.Websites.Unavailable":"Website down","Measurement.SummaryText.Websites.Accessible":"On {date}, {WebsiteURL} was accessible when tested on {network} in {country}.","Measurement.SummaryText.Websites.Anomaly":"On {date}, {WebsiteURL} presented signs of {reason} on {network} in {country}.\n\nThis might mean that {WebsiteURL} was blocked, but <link-to-docs>false positives</link-to-docs> can occur.\n\nPlease explore the network measurement data below.","Measurement.SummaryText.Websites.Anomaly.BlockingReason.DNS":"DNS tampering","Measurement.SummaryText.Websites.Anomaly.BlockingReason.TCP":"TCP/IP blocking","Measurement.SummaryText.Websites.Anomaly.BlockingReason.HTTP-failure":"HTTP blocking (HTTP requests failed)","Measurement.SummaryText.Websites.Anomaly.BlockingReason.HTTP-diff":"HTTP blocking (a blockpage might be served)","Measurement.SummaryText.Websites.ConfirmedBlocked":"On {date}, {WebsiteURL} was blocked on {network} in {country}.\n\nThis is confirmed because a block page was served, as illustrated through the network measurement data below.","Measurement.SummaryText.Websites.Failed":"On {date}, the test for {WebsiteURL} failed on {network} in {country}.","Measurement.SummaryText.Websites.Down":"On {date}, {WebsiteURL} was down on {network} in {country}.","Measurement.Details.Websites.Failures.Heading":"Failures","Measurement.Details.Websites.Failures.Label.HTTP":"HTTP Experiment","Measurement.Details.Websites.Failures.Label.DNS":"DNS Experiment","Measurement.Details.Websites.Failures.Label.Control":"Control","Measurement.Details.Websites.Failures.Values.Null":"null","Measurement.Details.Websites.DNSQueries.Heading":"DNS Queries","Measurement.Details.Websites.DNSQueries.Label.Resolver":"Resolver","Measurement.Details.Websites.TCP.Heading":"TCP Connections","Measurement.Details.Websites.TCP.ConnectionTo":"Connection to {destination} {connectionStatus}.","Measurement.Details.Websites.TCP.ConnectionTo.Success":"succeeded","Measurement.Details.Websites.TCP.ConnectionTo.Failed":"failed","Measurement.Details.Websites.TCP.ConnectionTo.Blocked":"was blocked","Measurement.Details.Websites.HTTP.Heading":"HTTP Requests","Measurement.Details.Websites.HTTP.Label.Response":"Response","Measurement.Details.Websites.HTTP.Request.URL":"URL","Measurement.Details.Websites.HTTP.Response.Body":"Response Body","Measurement.Details.Websites.HTTP.Response.Headers":"Response Headers","Measurement.CommonDetails.Label.MsmtID":"Report ID","Measurement.CommonDetails.Label.Platform":"Platform","Measurement.CommonDetails.Label.Software":"Software Name","Measurement.CommonDetails.Label.Engine":"Measurement Engine","Measurement.CommonDetails.Value.Unavailable":"Unavailable","Measurement.CommonDetails.RawMeasurement.Heading":"Raw Measurement Data","Measurement.CommonDetails.RawMeasurement.Download":"Download JSON","Measurement.CommonDetails.RawMeasurement.Unavailable":"Unavailable","Measurement.CommonDetails.RawMeasurement.Expand":"Expand All","Measurement.CommonDetails.Label.Resolver":"Resolver","Measurement.CommonDetails.Label.ResolverASN":"Resolver ASN","Measurement.CommonDetails.Label.ResolverIP":"Resolver IP","Measurement.CommonDetails.Label.ResolverNetworkName":"Resolver Network Name","Measurement.CommonDetails.Label.UserFeedback":"User Feedback","Measurement.Hero.Status.NDT.Title":"Results","Measurement.Status.Info.Label.Download":"Download","Measurement.Status.Info.Label.Upload":"Upload","Measurement.Status.Info.Label.Ping":"Ping","Measurement.Status.Info.Label.Server":"Server","Measurement.Status.Info.NDT.Error":"Failed Test","Measurement.Details.Performance.Heading":"Performance Details","Measurement.Details.Performance.Label.AvgPing":"Average Ping","Measurement.Details.Performance.Label.MaxPing":"Max Ping","Measurement.Details.Performance.Label.MSS":"MSS","Measurement.Details.Performance.Label.RetransmitRate":"Retransmission Rate","Measurement.Details.Performance.Label.PktLoss":"Packet Loss","Measurement.Details.Performance.Label.OutOfOrder":"Out of Order","Measurement.Details.Performance.Label.Timeouts":"Timeouts","Measurement.Hero.Status.Dash.Title":"Results","Measurement.Status.Info.Label.VideoQuality":"Video Quality","Measurement.Status.Info.Label.Bitrate":"Median Bitrate","Measurement.Status.Info.Label.Delay":"Playout Delay","Measurement.Status.Hint.Telegram.Blocked":"Telegram is likely blocked","Measurement.Status.Hint.Telegram.Reachable":"Telegram is accessible","Measurement.Status.Hint.Telegram.Failed":"The Telegram test failed","Measurement.Details.SummaryText.Telegram.Reachable":"On {date}, Telegram was reachable on {network} in {country}. \n\nOONI's Telegram test successfully connected to Telegram's endpoints and web interface (web.telegram.org).","Measurement.Details.SummaryText.Telegram.AppFailure":"On {date}, the testing of Telegram's mobile app presented signs of blocking on {network} in {country}.\n\nThis might mean that Telegram's mobile app was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.Telegram.DesktopFailure":"On {date}, the testing of Telegram's web interface (web.telegram.org) presented signs of blocking on {network} in {country}.\n\nThis might mean that web.telegram.org was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.Telegram.DesktopAndAppFailure":"On {date}, the testing of Telegram's mobile app and web interface (web.telegram.org) presented signs of blocking on {network} in {country}.\n\nThis might mean that both Telegram's mobile app and web interface were blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.Telegram.Endpoint.Label.Mobile":"Mobile App","Measurement.Details.Telegram.Endpoint.Label.Web":"Telegram Web","Measurement.Details.Endpoint.Status.Unknown":"Unknown","Measurement.Details.Telegram.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.Telegram.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.Telegram.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Details.Hint.WhatsApp.Reachable":"WhatsApp is accessible","Measurement.Status.Hint.WhatsApp.Blocked":"WhatsApp is likely blocked","Measurement.Status.Hint.WhatsApp.Failed":"The WhatsApp test failed","Measurement.Details.SummaryText.WhatsApp.Reachable":"On {date}, WhatsApp was reachable on {network} in {country}. \n\nOONI's WhatsApp test successfully connected to WhatsApp's endpoints, registration service and web interface (web.whatsapp.com).","Measurement.Details.SummaryText.WhatsApp.AppFailure":"On {date}, the testing of WhatsApp's mobile app presented signs of blocking on {network} in {country}.\n\nThis might mean that WhatsApp's mobile app was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.WhatsApp.DesktopFailure":"On {date}, the testing of WhatsApp's web interface (web.whatsapp.com) presented signs of blocking on {network} in {country}.\n\nThis might mean that web.whatsapp.com was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.WhatsApp.DesktopAndAppFailure":"On {date}, the testing of WhatsApp's mobile app and web interface (web.whatsapp.com) presented signs of blocking on {network} in {country}.\n\nThis might mean that both WhatsApp's mobile app and web interface were blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.WhatsApp.Endpoint.Label.Mobile":"Mobile App","Measurement.Details.WhatsApp.Endpoint.Label.Web":"WhatsApp Web","Measurement.Details.WhatsApp.Endpoint.Label.Registration":"Registration","Measurement.Details.WhatsApp.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.WhatsApp.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.WhatsApp.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Status.Hint.FacebookMessenger.Reachable":"Facebook Messenger is accessible","Measurement.Status.Hint.FacebookMessenger.Blocked":"Facebook Messenger is likely blocked","Measurement.Status.Hint.FacebookMessenger.Failed":"The Facebook Messenger test failed","Measurement.Details.SummaryText.FacebookMessenger.Reachable":"On {date}, Facebook Messenger was reachable on {network} in {country}.","Measurement.Details.SummaryText.FacebookMessenger.TCPFailure":"TCP connections to Facebook's endpoints failed.","Measurement.Details.SummaryText.FacebookMessenger.DNSFailure":"DNS lookups did not resolve to Facebook IP addresses.","Measurement.Details.SummaryText.FacebookMessenger.DNSSuccess":"DNS lookups resolved to Facebook IP addresses.","Measurement.Details.SummaryText.FacebookMessenger.TCPSuccess":"TCP connections to Facebook's enpoints succeeded.","Measurement.Details.FacebookMessenger.TCP.Label.Title":"TCP connections","Measurement.Details.FacebookMessenger.DNS.Label.Title":"DNS lookups","Measurement.Details.FacebookMessenger.TCPFailed":"TCP connections failed","Measurement.Details.FacebookMessenger.DNSFailed":"DNS lookups failed","Measurement.Details.FacebookMessenger.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.FacebookMessenger.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.FacebookMessenger.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Status.Hint.Signal.Blocked":"Signal is likely blocked","Measurement.Status.Hint.Signal.Reachable":"Signal is accessible","Measurement.Status.Hint.Signal.Failed":"The Signal test failed","Measurement.Details.SummaryText.Signal.Reachable":"On {date}, [Signal](https://signal.org/) was reachable on {network} in {country}. \n\nThe [OONI Probe Signal test](https://ooni.org/nettest/signal) successfully connected to Signal's endpoints.","Measurement.Details.SummaryText.Signal.Blocked":"On {date}, the testing of the [Signal app](https://signal.org/) presented signs of blocking on {network} in {country}.\n\nThis might mean that Signal was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other Signal measurements from the same network during the same time period (if they're available).","Measurement.Hero.Status.HTTPHeaderManipulation.NoMiddleBoxes":"No middleboxes detected","Measurement.HTTPHeaderManipulation.NoMiddleBoxes.SummaryText":"On {date}, no network anomaly was detected on {network} in {country} when communicating with our servers.","Measurement.Hero.Status.HTTPHeaderManipulation.MiddleboxesDetected":"Network tampering","Measurement.HTTPHeaderManipulation.MiddleBoxesDetected.SummaryText":"On {date}, network traffic was manipulated when contacting our control servers. \n\nThis means that there might be a middlebox on {network} in {country}, which could be responsible for censorship and/or surveillance.","Measurement.Hero.Status.HTTPInvalidReqLine.NoMiddleBoxes":"No middleboxes detected","Measurement.HTTPInvalidReqLine.NoMiddleBoxes.SummaryText":"On {date}, no network anomaly was detected on {network} in {country} when communicating with our servers.","Measurement.Hero.Status.HTTPInvalidReqLine.MiddleboxesDetected":"Network tampering","Measurement.HTTPInvalidReqLine.MiddleboxesDetected.SummaryText":"On {date}, network traffic was manipulated when contacting our control servers. \n\nThis means that there might be a middlebox on {network} in {country}, which could be responsible for censorship and/or surveillance.","Measurement.HTTPInvalidReqLine.YouSent":"You Sent","Measurement.HTTPInvalidReqLine.YouReceived":"You Received","Measurement.Hero.Status.TorVanilla.Blocked":"Tor is likely blocked","Measurement.Hero.Status.TorVanilla.Reachable":"Tor is accessible","Measurement.Details.SummaryText.TorVanilla.Blocked":"On {date}, OONI's Vanilla Tor test did not manage to bootstrap a connection to the [Tor network](https://www.torproject.org/).\n\nThis might mean that access to the Tor network was blocked on {network} in {country}, but [false positives can occur](https://ooni.org/support/faq/#why-do-false-positives-occur).\n\nPlease explore the network measurement data below. Check other Tor measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.TorVanilla.Reachable":"OONI's Vanilla Tor test successfully bootstrapped a connection to the [Tor network](https://www.torproject.org/).\n\nThis means that the Tor network was reachable from {network} in {country} on {date}.","Measurement.Details.VanillaTor.Endpoint.Label.Reachability":"Reachability","Measurement.Status.Hint.Psiphon.Reachable":"Psiphon works","Measurement.Status.Hint.Psiphon.Blocked":"Psiphon is likely blocked","Measurement.Status.Hint.Psiphon.BootstrappingError":"Psiphon is likely blocked (bootstrap error)","Measurement.Details.SummaryText.Psiphon.OK":"On {date}, [Psiphon](https://psiphon.ca/) worked on {network} in {country}.\n\nThe [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was able to successfully bootstrap Psiphon and ensure that the app works.","Measurement.Details.SummaryText.Psiphon.Blocked":"On {date}, [Psiphon](https://psiphon.ca/) did not appear to work on {network} in {country}.\n\nWhile the [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was able to bootstrap Psiphon, it was unable to fetch a webpage from the internet. \n\nThis suggests that the Psiphon app may have been blocked on this network. \n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. Please explore the network measurement data below and compare it with other relevant measurements (if they're available).","Measurement.Details.SummaryText.Psiphon.BootstrappingError":"On {date}, [Psiphon](https://psiphon.ca/) did not work on {network} in {country}.\n\nThe [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was unable to bootstrap Psiphon.\n\nThis suggests that the Psiphon app may have been blocked on this network.\n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. Please explore the network measurement data below and compare it with other relevant measurements (if they're available).","Measurement.Details.Psiphon.BootstrapTime.Label":"Bootstrap Time","Measurement.Status.Hint.Tor.Reachable":"Tor works","Measurement.Status.Hint.Tor.Blocked":"Tor is likely blocked","Measurement.Status.Hint.Tor.Error":"Tor test failed","Measurement.Details.SummaryText.Tor.OK":"On {date}, [Tor](https://www.torproject.org/) worked on {network} in {country}.\n\nAs part of [OONI Probe Tor testing](https://ooni.org/nettest/tor/), all reachability measurements of selected Tor directory authorities and bridges were successful.","Measurement.Details.SummaryText.Tor.Blocked":"On {date}, [Tor](https://www.torproject.org/) did not appear to work on {network} in {country}.\n\nThe [OONI Probe Tor test](https://ooni.org/nettest/tor/) failed in performing certain measurements. More details are available through the network measurement data provided below.\n\nThis suggests that Tor may have been blocked on this network.\n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.","Measurement.Details.SummaryText.Tor.Error":"On {date}, the Tor test failed on {network} in {country}.","Measurement.Details.Tor.Bridges.Label.Title":"Tor Browser Bridges","Measurement.Details.Tor.Bridges.Label.OK":"{bridgesAccessible}/{bridgesTotal} OK","Measurement.Details.Tor.DirAuth.Label.Title":"Tor Directory Authorities","Measurement.Details.Tor.DirAuth.Label.OK":"{dirAuthAccessible}/{dirAuthTotal} OK","Measurement.Details.Tor.Table.Header.Name":"Name","Measurement.Details.Tor.Table.Header.Address":"Address","Measurement.Details.Tor.Table.Header.Type":"Type","Measurement.Status.Hint.TorSnowflake.Reachable":"Tor Snowflake works","Measurement.Status.Hint.TorSnowflake.Blocked":"Tor Snowflake does not work","Measurement.Status.Hint.TorSnowflake.Error":"Tor Snowflake test failed","Measurement.Details.SummaryText.TorSnowflake.OK":"On {date}, [Tor Snowflake](https://www.torproject.org/) worked on {network} in {country}.\n\nThe [OONI Probe Tor Snowflake test](https://ooni.org/nettest/torsf/) was able to successfully bootstrap Snowflake.","Measurement.Details.SummaryText.TorSnowflake.Blocked":"On {date}, [Tor Snowflake](https://www.torproject.org/) does not work on {network} in {country}.\n\nThe [OONI Probe Tor Snowflake test](https://ooni.org/nettest/torsf/) failed to bootstrap Snowflake.","Measurement.Details.SummaryText.TorSnowflake.Error":"On {date}, the Tor Snowflake test failed on {network} in {country}.","Measurement.Details.TorSnowflake.BootstrapTime.Label":"Bootstrap Time","Measurement.Details.TorSnowflake.Error.Label":"Failure","Measurement.Metadata.TorSnowflake.Reachable":"Tor Snowflake was reachable in {country}","Measurement.Metadata.TorSnowflake.UnReachable":"Tor Snowflake was NOT reachable in {country}","Measurement.Metadata.TorSnowflake.Error":"Tor Snowflake test failed in {country}","Measurement.Status.Hint.RiseupVPN.Reachable":"RiseupVPN works","Measurement.Status.Hint.RiseupVPN.Blocked":"RiseupVPN is likely blocked","Measurement.Status.Hint.RiseupVPN.Failed":"The RiseupVPN test failed","Measurement.Details.SummaryText.RiseupVPN.OK":"On {date}, [RiseupVPN](https://riseup.net/vpn) was reachable on {network} in {country}. \n\nThe [OONI Probe RiseupVPN test](https://ooni.org/nettest/riseupvpn/) successfully connected to RiseupVPN's bootstrap servers and gateways.","Measurement.Details.SummaryText.RiseupVPN.Blocked":"On {date}, the testing of [RiseupVPN](https://riseup.net/vpn) presented signs of blocking on {network} in {country}.\n\nThis might mean that RiseupVPN was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other RiseupVPN measurements from the same network during the same time period (if they're available).","Measurement.Metadata.RiseupVPN.Reachable":"RiseupVPN was reachable in {country}","Measurement.Metadata.RiseupVPN.Blocked":"RiseupVPN was not reachable in {country}","Measurement.Hero.Status.Default":"Measurement Report","Measurement.Feedback.Title":"Verify the measurement","Measurement.Feedback.Description":"Please share feedback based on what you see in the measurement data.","Measurement.Feedback.ok":"It's OK","Measurement.Feedback.down":"It's down","Measurement.Feedback.down.unreachable":"It's unreachable","Measurement.Feedback.down.misconfigured":"It's misconfigured","Measurement.Feedback.blocked":"It's blocked","Measurement.Feedback.blocked.tcp":"TCP/IP blocking","Measurement.Feedback.blocked.tls":"TLS blocking","Measurement.Feedback.blocked.blockpage":"Block page","Measurement.Feedback.blocked.blockpage.http":"HTTP block page","Measurement.Feedback.blocked.blockpage.dns":"DNS block page","Measurement.Feedback.blocked.blockpage.server_side":"Server-side block page","Measurement.Feedback.blocked.blockpage.server_side.captcha":"CAPTCHA","Measurement.Feedback.blocked.dns":"DNS blocking without block page","Measurement.Feedback.blocked.dns.inconsistent":"Inconsistent DNS response","Measurement.Feedback.blocked.dns.nxdomain":"NXDOMAIN error","Measurement.Feedback.Success.Title":"Thank you!","Measurement.Feedback.Success.Description":"Your feedback will help improve OONI measurements.","Measurement.Feedback.Failure":"Something went wrong, please try again.","Measurement.Feedback.Login.Title":"Please log in to continue","Measurement.Feedback.Login.Confirmation.Title":"Login link sent","Measurement.Feedback.Login.Confirmation.Text":"Please check your email for a link to log in to your account.","Measurement.Feedback.Login.Description":"We will send a link to your email. We do not store emails.","Measurement.Feedback.ExistingFeedback":"Your previous feedback","Navbar.Search":"Search","Navbar.Countries":"Countries","Navbar.Charts.Circumvention":"Circumvention Charts","Navbar.Charts.MAT":"MAT Charts","Network.Summary.TotalMeasurements":"Total number of measurements: **{measurementsTotal}**","Network.Summary.FirstMeasurement":"Date of the first measurement: **{formattedDate}**","Network.Summary.Countries":"Network observed in countries:","Network.Summary.Country.Measurements":"({measurementsTotal} measurements)","Network.NoData.Title":"Let's collect more data!","Network.NoData.Text":"We don't have enough data for this network to show the charts. Please run OONI Probe to collect more measurements.","Footer.Text.Slogan":"Global community measuring internet censorship around the world.","Footer.Heading.About":"About","Footer.Heading.OONIProbe":"OONI Probe","Footer.Heading.Updates":"Updates","Footer.Link.About":"OONI","Footer.Link.DataPolicy":"Data Policy","Footer.Link.DataLicense":"Data License","Footer.Link.Contact":"Contact","Footer.Link.Probe":"Install","Footer.Link.Tests":"Tests","Footer.Link.Code":"Source code","Footer.Link.API":"API","Footer.Link.Blog":"Blog","Footer.Link.Twitter":"Twitter","Footer.Link.MailingList":"Mailing list","Footer.Link.Slack":"Slack","Footer.Text.Copyright":"© {currentYear} Open Observatory of Network Interference (OONI)","Footer.Text.CCommons":"Content available under a Creative Commons license.","Footer.Text.Version":"Version","CategoryCode.ALDR.Name":"Alcohol & Drugs","CategoryCode.REL.Name":"Religion","CategoryCode.PORN.Name":"Pornography","CategoryCode.PROV.Name":"Provocative Attire","CategoryCode.POLR.Name":"Political Criticism","CategoryCode.HUMR.Name":"Human Rights Issues","CategoryCode.ENV.Name":"Environment","CategoryCode.MILX.Name":"Terrorism and Militants","CategoryCode.HATE.Name":"Hate Speech","CategoryCode.NEWS.Name":"News Media","CategoryCode.XED.Name":"Sex Education","CategoryCode.PUBH.Name":"Public Health","CategoryCode.GMB.Name":"Gambling","CategoryCode.ANON.Name":"Anonymization and circumvention tools","CategoryCode.DATE.Name":"Online Dating","CategoryCode.GRP.Name":"Social Networking","CategoryCode.LGBT.Name":"LGBTQ+","CategoryCode.FILE.Name":"File-sharing","CategoryCode.HACK.Name":"Hacking Tools","CategoryCode.COMT.Name":"Communication Tools","CategoryCode.MMED.Name":"Media sharing","CategoryCode.HOST.Name":"Hosting and Blogging Platforms","CategoryCode.SRCH.Name":"Search Engines","CategoryCode.GAME.Name":"Gaming","CategoryCode.CULTR.Name":"Culture","CategoryCode.ECON.Name":"Economics","CategoryCode.GOVT.Name":"Government","CategoryCode.COMM.Name":"E-commerce","CategoryCode.CTRL.Name":"Control content","CategoryCode.IGO.Name":"Intergovernmental Organizations","CategoryCode.MISC.Name":"Miscellaneous content","CategoryCode.ALDR.Description":"Sites devoted to the use, paraphernalia, and sale of drugs and alcohol irrespective of the local legality.","CategoryCode.REL.Description":"Sites devoted to discussion of religious issues, both supportive and critical, as well as discussion of minority religious groups.","CategoryCode.PORN.Description":"Hard-core and soft-core pornography.","CategoryCode.PROV.Description":"Websites which show provocative attire and portray women in a sexual manner, wearing minimal clothing.","CategoryCode.POLR.Description":"Content that offers critical political viewpoints. Includes critical authors and bloggers, as well as oppositional political organizations. Includes pro-democracy content, anti-corruption content as well as content calling for changes in leadership, governance issues, legal reform, etc.","CategoryCode.HUMR.Description":"Sites dedicated to discussing human rights issues in various forms. Includes women's rights and rights of minority ethnic groups.","CategoryCode.ENV.Description":"Pollution, international environmental treaties, deforestation, environmental justice, disasters, etc.","CategoryCode.MILX.Description":"Sites promoting terrorism, violent militant or separatist movements.","CategoryCode.HATE.Description":"Content that disparages particular groups or persons based on race, sex, sexuality or other characteristics.","CategoryCode.NEWS.Description":"This category includes major news outlets (BBC, CNN, etc.) as well as regional news outlets and independent media.","CategoryCode.XED.Description":"Includes contraception, abstinence, STDs, healthy sexuality, teen pregnancy, rape prevention, abortion, sexual rights, and sexual health services.","CategoryCode.PUBH.Description":"HIV, SARS, bird flu, centers for disease control, World Health Organization, etc.","CategoryCode.GMB.Description":"Online gambling sites. Includes casino games, sports betting, etc.","CategoryCode.ANON.Description":"Sites that provide tools used for anonymization, circumvention, proxy-services and encryption.","CategoryCode.DATE.Description":"Online dating services which can be used to meet people, post profiles, chat, etc.","CategoryCode.GRP.Description":"Social networking tools and platforms.","CategoryCode.LGBT.Description":"A range of gay-lesbian-bisexual-transgender queer issues. (Excluding pornography)","CategoryCode.FILE.Description":"Sites and tools used to share files, including cloud-based file storage, torrents and P2P file-sharing tools.","CategoryCode.HACK.Description":"Sites dedicated to computer security, including news and tools. Includes malicious and non-malicious content.","CategoryCode.COMT.Description":"Sites and tools for individual and group communications. Includes webmail, VoIP, instant messaging, chat and mobile messaging applications.","CategoryCode.MMED.Description":"Video, audio or photo sharing platforms.","CategoryCode.HOST.Description":"Web hosting services, blogging and other online publishing platforms.","CategoryCode.SRCH.Description":"Search engines and portals.","CategoryCode.GAME.Description":"Online games and gaming platforms, excluding gambling sites.","CategoryCode.CULTR.Description":"Content relating to entertainment, history, literature, music, film, books, satire and humour.","CategoryCode.ECON.Description":"General economic development and poverty related topics, agencies and funding opportunities.","CategoryCode.GOVT.Description":"Government-run websites, including military sites.","CategoryCode.COMM.Description":"Websites of commercial services and products.","CategoryCode.CTRL.Description":"Benign or innocuous content used as a control.","CategoryCode.IGO.Description":"Websites of intergovernmental organizations such as the United Nations.","CategoryCode.MISC.Description":"Sites that don't fit in any category. (XXX Things in here should be categorised)","Country.Heading.Overview":"Overview","Country.Heading.Websites":"Websites","Country.Heading.Apps":"Apps","Country.Heading.Shutdowns":"Shutdowns","Country.Heading.NetworkProperties":"Networks","Country.Overview.Heading.NwInterference":"In a nutshell","Country.Overview.NwInterference.Middleboxes.Blocked":"Middleboxes were detected on {middleboxCount} network(s)","Country.Overview.NwInterference.Middleboxes.Normal":"No middleboxes were detected on tested networks","Country.Overview.NwInterference.Middleboxes.NoData":"Not enough data available on middleboxes","Country.Overview.NwInterference.IM.Blocked":"Instant messaging apps were likely blocked","Country.Overview.NwInterference.IM.Normal":"No instant messaging apps were blocked on tested networks","Country.Overview.NwInterference.IM.NoData":"Not enough data available on instant messaging apps","Country.Overview.NwInterference.CircumventionTools.Blocked":"Circumvention tools were likely blocked","Country.Overview.NwInterference.CircumventionTools.Normal":"No circumvention tools were blocked on tested networks","Country.Overview.NwInterference.CircumventionTools.NoData":"Not enough data available on circumvention tools","Country.Overview.NwInterference.Websites.Blocked":"OONI data confirms the blocking of websites","Country.Overview.NwInterference.Websites.Normal":"The blocking of websites is not confirmed","Country.Overview.NwInterference.Websites.NoData":"Not enough data available on blocked websites","Country.Overview.Heading.TestsByClass":"Measurement Coverage","Country.Overview.Heading.TestsByClass.Description":"The graph below provides an overview of OONI Probe measurement coverage. It shows how many results have been collected from each OONI Probe test category, as well as how many networks have been covered by tests. \n\nBy looking at this graph, you can understand if there is enough data to draw meaningful conclusions. If there is not enough data and you are in the country in question, [install OONI Probe](https://ooni.org/install), run tests, and contribute data!","Country.Overview.TestsByClass.Websites":"Websites","Country.Overview.TestsByClass.InstantMessaging":"Instant Messaging","Country.Overview.TestsByClass.Performance":"Performance","Country.Overview.TestsByClass.Middleboxes":"Middleboxes","Country.Overview.TestsByClass.Circumvention":"Circumvention Tools","Country.Overview.FeaturedResearch":"Research Reports","Country.Overview.FeaturedResearch.None":"We haven't published a research report based on OONI data from this country yet. \n\nWe encourage you to use OONI data for your research!","Country.Overview.SummaryTextTemplate":"OONI Probe users in **{countryName}** have collected [**{measurementCount}** measurements]({linkToMeasurements}) from **{networkCovered}** local networks.\n\nExplore the data below to check the accessibility and/or blocking of sites and services.","Country.Overview.NoData.Title":"Let's collect more data!","Country.Overview.NoData.CallToAction":"We don’t have enough measurements for **{country}** to show a chart. If you are in {country} or know people there, tell them to run OONI Probe to collect more measurements.","Country.Overview.NoData.Button.InstallProbe":"Install OONI Probe","Country.Overview.NoData.Button.OoniRunLink":"Create OONI Run Link","Country.Meta.Title":"Internet Censorship in {countryName} - OONI Explorer","Country.Meta.Description":"OONI Probe users in {countryName} have collected {measurementCount} measurements from {networkCount} local networks. Explore the data on OONI Explorer.","Country.PeriodFilter.Label":"Show results from","Country.PeriodFilter.Option.30Days":"Last 30 Days","Country.PeriodFilter.Option.2Months":"Last 2 Months","Country.PeriodFilter.Option.3Months":"Last 3 Months","Country.PeriodFilter.Option.6Months":"Last 6 Months","Country.Websites.Description":"Check whether websites have been blocked.\n\nTesting methodology: OONI's [Web Connectivity test](https://ooni.org/nettest/web-connectivity/), designed to measure the DNS, HTTP, and TCP/IP blocking of websites. \n\nTested websites: [Citizen Lab test lists](https://github.com/citizenlab/test-lists)\n\nIf you'd like to see results on the testing of different websites, please [contribute to test lists](https://ooni.org/get-involved/contribute-test-lists/) or test the sites of your choice via the [OONI Probe mobile app](https://ooni.org/install/). \n\nPlease note that unless a block page is served, some anomalous measurements may contain [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur). We therefore encourage you to examine anomalous measurements in depth and over time.","Country.Websites.Description.MoreLinkText":"","Country.Websites.Heading.BlockedByCategory":"Categories of Blocked Websites","Country.Websites.BlockedByCategory.Description":"Websites that fall under the following categories were blocked on the {selectedASN} network.","Country.Websites.TestedWebsitesCount":"URLs tested","Country.Websites.Labels.ResultsPerPage":"Results per page","Country.Websites.URLSearch.Placeholder":"Search for URL","Country.Websites.URLCharts.Legend.Label.Blocked":"Confirmed Blocked","Country.Websites.URLCharts.ExploreMoreMeasurements":"Explore more measurements","Country.Websites.URLCharts.Pagination.Previous":"Previous Page","Country.Websites.URLCharts.Pagination.Next":"Next Page","Country.Apps.Description":"Check whether instant messaging apps and circumvention tools are blocked.\n\nThe following results were collected through the use of [OONI Probe tests](https://ooni.org/nettest/) designed to measure the blocking of WhatsApp, Facebook Messenger, Telegram and Signal. \n\nWe also share results on the testing of circumvention tools, like [Tor](https://www.torproject.org/) and [Psiphon](https://psiphon.ca/).","Country.Apps.Label.LastTested":"Last tested","Country.Apps.Label.TestedNetworks":"tested networks","Country.Apps.Button.ShowMore":"Show More","Country.Shutdowns":"Internet shutdowns","Country.Shutdowns.Description":"Monitor Internet shutdowns through public, third-party data sources. \n\nThe following graph shares data from Georgia Tech's [Internet Outage Detection and Analysis (IODA)](https://ioda.inetintel.cc.gatech.edu/) project, [Google Transparency Reports (Google traffic data)](https://transparencyreport.google.com/traffic/overview?hl=en), and [Cloudflare Radar](https://radar.cloudflare.com/).\n\nIf you notice a drop in all signals, that might serve as an indicator of an Internet shutdown. \n\nLearn more about Internet shutdowns and their impact through the [#KeepItOn campaign](https://www.accessnow.org/keepiton/).","Country.NetworkProperties.Description":"Check the speed and performance of networks.\n\nThe following results were collected through the use of [OONI Probe's performance and middlebox tests](https://ooni.org/nettest/). You can check the speed and performance of tested networks, as well as video streaming performance. \n\nYou can also learn whether middleboxes were detected in tested networks. Middleboxes are network appliances that can be used for a variety of networking purposes (such as caching), but sometimes they're used to implement internet censorship and/or surveillance.","Country.NetworkProperties.Heading.Summary":"Summary","Country.NetworkProperties.Heading.Networks":"Networks","Country.NetworkProperties.InfoBox.Label.AverageDownload":"Average Download","Country.NetworkProperties.InfoBox.Label.AverageUpload":"Average Upload","Country.NetworkProperties.InfoBox.Label.Covered":"Covered","Country.NetworkProperties.InfoBox.Label.Middleboxes":"Middleboxes detected","Country.NetworkProperties.InfoBox.Units.Mbits":"Mbit/s","Country.NetworkProperties.InfoBox.Units.Networks.Singular":"Network","Country.NetworkProperties.InfoBox.Units.Networks.Plural":"Networks","Country.NetworkProperties.InfoBox.Label.AverageStreaming":"Average Streaming","Country.NetworkProperties.InfoBox.Label.AveragePing":"Average Ping","Country.NetworkProperties.InfoBox.Units.Milliseconds":"ms","Country.NetworkProperties.InfoBox.Label.Middleboxes.Found":"Middleboxes detected","Country.NetworkProperties.InfoBox.Label.Middleboxes.NotFound":"No middleboxes detected","Country.NetworkProperties.Button.ShowMore":"Show more networks","Country.Label.NoData":"No Data Available","Search.PageTitle":"Search through millions of Internet censorship measurements","Search.Sidebar.Domain":"Domain","Search.Sidebar.Domain.Placeholder":"e.g. twitter.com or 1.1.1.1","Search.Sidebar.Domain.Error":"Please enter a valid domain name or IP address, such as twitter.com or 1.1.1.1","Search.Sidebar.Input":"Input","Search.Sidebar.Input.Placeholder":"e.g., https://twitter.com/","Search.Sidebar.Input.Error":"Please enter a full URL (e.g. https://twitter.com/) with a trailing slash (`/`) or an IP address","Search.Sidebar.Categories":"Website Categories","Search.Sidebar.Categories.All":"Any","Search.Sidebar.Status":"Status","Search.Sidebar.TestName":"Test Name","Search.Sidebar.TestName.AllTests":"Any","Search.Sidebar.Country":"Country","Search.Sidebar.Country.AllCountries":"Any","Search.Sidebar.ASN":"ASN","Search.Sidebar.ASN.example":"e.g. AS30722","Search.Sidebar.ASN.Error":"Valid formats: AS1234, 1234","Search.Sidebar.From":"From","Search.Sidebar.Until":"Until","Search.Sidebar.HideFailed":"Hide failed measurements","Search.Sidebar.Button.FilterResults":"Filter Results","Search.FilterButton.AllResults":"All Results","Search.FilterButton.Confirmed":"Confirmed","Search.FilterButton.Anomalies":"Anomalies","Search.FilterButton.Search":"Search","Search.Filter.SortBy":"Sort by","Search.Filter.SortBy.Date":"Date","Search.WebConnectivity.Results.Blocked":"Confirmed","Search.HTTPRequests.Results.Anomaly":"","Search.HTTPRequests.Results.Blocked":"","Search.HTTPRequests.Results.Error":"","Search.HTTPRequests.Results.Reachable":"","Search.NDT.Results":"","Search.DASH.Results":"","Search.VanillaTor.Results":"","Search.BridgeReachability.Results":"","Search.LegacyTests.Results":"","Search.Results.Empty.Heading":"No Results Found","Search.Results.Empty.Description":"Please try changing the filters to get better results.","Search.Button.LoadMore":"Load more","Search.Error.Message":"This query took too long to complete. Please try adjusting the search filters or view the example queries in the [Highlights section of the homepage](/#highlights).\n\nIf you are interested in using OONI data in batch, we recommend the [ooni-data Amazon S3 bucket](https://ooni.org/post/mining-ooni-data/) or the [aggregation API](https://api.ooni.io/apidocs/#/default/get_api_v1_aggregation).\n\nWe are working on improving the performance of OONI Explorer. To track our work on this, [see the open issues on the ooni/api repository](https://github.com/ooni/api/issues?q=is%3Aissue+is%3Aopen+label%3Aoptimization).","Search.Error.Details.Label":"Server Response","Home.Banner.Title.UncoverEvidence":"Uncover evidence of internet censorship worldwide","Home.Banner.Subtitle.ExploreCensorshipEvents":"Open data collected by the global OONI community","Home.Banner.Button.Explore":"Explore","Home.Banner.Stats.Measurements":"Measurements","Home.Banner.Stats.Countries":"Countries","Home.Banner.Stats.Networks":"Networks","Home.About.SummaryText":"OONI Explorer is an open data resource on internet censorship around the world. \n\nSince 2012, millions of network measurements have been collected from more than 200 countries. OONI Explorer sheds light on internet censorship and other forms of network interference worldwide.\n\nTo contribute to this open dataset, [install OONI Probe](https://ooni.org/install/) and run tests!","Home.Websites&Apps.Title":"Blocking of Websites & Apps","Home.Websites&Apps.SummaryText":"Discover blocked websites around the world. Check whether WhatsApp, Facebook Messenger, and Telegram are blocked.","Home.Search&Filter.Title":"Search","Home.Search&Filter.SummaryText":"Explore OONI measurements with a powerful search tool. View the most recently blocked websites. Compare internet censorship across networks.","Home.NetworkProperties.Title":"Network Performance","Home.NetworkProperties.SummaryText":"Check the speed and performance of thousands of networks around the world. Explore data on video streaming performance.","Home.MonthlyStats.Title":"Monthly coverage worldwide","Home.MonthlyStats.SummaryText":"OONI Explorer hosts millions of network measurements collected from more than 200 countries since 2012. Every day, OONI Explorer is updated with new measurements!\n\nLast month, {measurementCount} OONI Probe measurements were collected from {networkCount} networks in {countryCount} countries. Explore the monthly usage of [OONI Probe](https://ooni.org/install/) through the stats below.","Home.Highlights.CTA":"We encourage you to <link-to-search>explore OONI measurements</link-to-search> to find more highlights!","Home.Highlights.Title":"Highlights","Home.Highlights.Description":"What can you learn from OONI Explorer? \n\nBelow we share some stories from [research reports](https://ooni.org/post/) based on OONI data.\n\nWe share these case studies to demonstrate how OONI's openly available data can be used and what types of stories can be told. \n\nWe encourage you to explore OONI data, discover more censorship cases, and to use OONI data as part of your research and/or advocacy.","Home.Highlights.Political":"Censorship during political events","Home.Highlights.Political.Description":"Internet censorship sometimes occurs in response to or in anticipation of political events, such as elections, protests, and riots. Below we share a few cases detected via OONI data and correlated with political events.","Home.Highlights.Media":"Media censorship","Home.Highlights.Media.Description":"Press freedom is threatened in countries that experience the blocking of media websites. Below we share a few cases detected through OONI data.","Home.Highlights.LGBTQI":"Blocking of LGBTQI sites","Home.Highlights.LGBTQI.Description":"Minority group sites are blocked around the world. Below we share a few cases on the blocking of LGBTQI sites.","Home.Highlights.Changes":"Censorship changes","Home.Highlights.Changes.Description":"OONI measurements have been collected on a continuous basis since 2012, enabling the identification of censorship changes around the world. Some examples include:","Home.Meta.Description":"OONI Explorer is an open data resource on Internet censorship around the world consisting of millions of measurements on network interference.","Home.Highlights.Explore":"Explore","Home.Highlights.ReadReport":"Read report","Countries.Heading.JumpToContinent":"Jump to continent","Countries.Search.NoCountriesFound":"No countries found with {searchTerm}","Countries.Search.Placeholder":"Search for countries","Countries.PageTitle":"Internet Censorship around the world","Error.404.PageNotFound":"Page Not Found","Error.404.GoBack":"Go back","Error.404.Heading":"The requested page does not exist","Error.404.Message":"We could not find the content you were looking for. Maybe try {measurmentLink} or look at {homePageLink}.","Error.404.MeasurmentLinkText":"exploring some measurement","Error.404.HomepageLinkText":"the homepage","MAT.Title":"OONI Measurement Aggregation Toolkit (MAT)","MAT.SubTitle":"Create charts based on aggregate views of real-time OONI data from around the world","MAT.JSONData":"JSON Data","MAT.CSVData":"CSV Data","MAT.Form.Label.XAxis":"Columns","MAT.Form.Label.YAxis":"Rows","MAT.Form.Label.TimeGrain":"Time Granularity","MAT.Form.Label.AxisOption.domain":"Domain","MAT.Form.Label.AxisOption.input":"Input","MAT.Form.Label.AxisOption.measurement_start_day":"Measurement Day","MAT.Form.Label.AxisOption.probe_cc":"Countries","MAT.Form.Label.AxisOption.category_code":"Website Categories","MAT.Form.Label.AxisOption.probe_asn":"ASN","MAT.Form.TimeGrainOption.hour":"Hour","MAT.Form.TimeGrainOption.day":"Day","MAT.Form.TimeGrainOption.week":"Week","MAT.Form.TimeGrainOption.month":"Month","MAT.Form.ConfirmationModal.Title":"Are you sure?","MAT.Form.ConfirmationModal.Message":"Duration too long. This can potentially slow down the page","MAT.Form.ConfirmationModal.No":"No","MAT.Form.ConfirmationModal.Button.Yes":"Yes","MAT.Form.Submit":"Show Chart","MAT.Form.All":"All","MAT.Form.AllCountries":"All Countries","MAT.Table.Header.ok_count":"OK Count","MAT.Table.Header.anomaly_count":"Anomaly Count","MAT.Table.Header.confirmed_count":"Confirmed Count","MAT.Table.Header.failure_count":"Failure Count","MAT.Table.Header.measurement_count":"Measurement Count","MAT.Table.Header.input":"URL","MAT.Table.Header.category_code":"Category Code","MAT.Table.Header.probe_cc":"Country","MAT.Table.Header.probe_asn":"ASN","MAT.Table.Header.blocking_type":"Blocking Type","MAT.Table.Header.domain":"Domain","MAT.Table.FilterPlaceholder":"Search {count} records…","MAT.Table.Search":"Search:","MAT.Table.Filters":"Filters","MAT.Charts.NoData.Title":"No Data Found","MAT.Charts.NoData.Description":"We are not able to produce a chart based on the selected filters. Please change the filters and try again.","MAT.Charts.NoData.Details":"Details:","MAT.Help.Box.Title":"Help","MAT.Help.Title":"FAQs","MAT.Help.Content":"# What is the MAT?\n\nOONI's Measurement Aggregation Toolkit (MAT) is a tool that enables you to generate your own custom charts based on **aggregate views of real-time OONI data** collected from around the world.\n\nOONI data consists of network measurements collected by [OONI Probe](https://ooni.org/install/) users around the world. \n\nThese measurements contain information about various types of **internet censorship**, such as the [blocking of websites and apps](https://ooni.org/nettest/) around the world. \n\n# Who is the MAT for?\n\nThe MAT was built for researchers, journalists, and human rights defenders interested in examining internet censorship around the world.\n\n# Why use the MAT?\n\nWhen examining cases of internet censorship, it's important to **look at many measurements at once** (\"in aggregate\") in order to answer key questions like the following:\n\n* Does the testing of a service (e.g. Facebook) present **signs of blocking every time that it is tested** in a country? This can be helpful for ruling out [false positives](https://ooni.org/support/faq/#what-are-false-positives).\n* What types of websites (e.g. human rights websites) are blocked in each country?\n* In which countries is a specific website (e.g. `bbc.com`) blocked?\n* How does the blocking of different apps (e.g. WhatsApp or Telegram) vary across countries?\n* How does the blocking of a service vary across countries and [ASNs](https://ooni.org/support/glossary/#asn)?\n* How does the blocking of a service change over time?\n\nWhen trying to answer questions like the above, we normally perform relevant data analysis (instead of inspecting measurements one by one). \n\nThe MAT incorporates our data analysis techniques, enabling you to answer such questions without any data analysis skills, and with the click of a button!\n\n# How to use the MAT?\n\nThrough the filters at the start of the page, select the parameters you care about in order to plot charts based on aggregate views of OONI data.\n\nThe MAT includes the following filters:\n\n* **Countries:** Select a country through the drop-down menu (the \"All Countries\" option will show global coverage)\n* **Test Name:** Select an [OONI Probe test](https://ooni.org/nettest/) based on which you would like to get measurements (for example, select `Web Connectivity` to view the testing of websites)\n* **Domain:** Type the domain for the website you would like to get measurements (e.g. `twitter.com`)\n* **Website categories:** Select the [website category](https://github.com/citizenlab/test-lists/blob/master/lists/00-LEGEND-new_category_codes.csv) for which you would like to get measurements (e.g. `News Media` for news media websites)\n* **ASN:** Type the [ASN](https://ooni.org/support/glossary/#asn) of the network for which you would like to get measurements (e.g. `AS30722` for Vodafone Italia)\n* **Date range:** Select the date range of the measurements by adjusting the `Since` and `Until` filters\n* **Columns:** Select the values that you would like to appear on the horizontal axis of your chart\n* **Rows:** Select the values that you would like to appear on the vertical axis of your chart\n\nDepending on what you would like to explore, adjust the MAT filters accordingly and click `Show Chart`. \n\nFor example, if you would like to check the testing of BBC in all countries around the world:\n\n* Type `www.bbc.com` under `Domain`\n* Select `Countries` under the `Rows`\n* Click `Show Chart`\n\nThis will plot numerous charts based on the OONI Probe testing of `www.bbc.com` worldwide.\n\n# Interpreting MAT charts\n\nThe MAT charts (and associated tables) include the following values:\n\n* **OK count:** Successful measurements (i.e. NO sign of internet censorship)\n* **Confirmed count:** Measurements from automatically **confirmed blocked websites** (e.g. a [block page](https://ooni.org/support/glossary/#block-page) was served)\n* **Anomaly count:** Measurements that provided **signs of potential blocking** (however, [false positives](https://ooni.org/support/faq/#what-are-false-positives) can occur) \n* **Failure count:** Failed experiments that should be discarded\n* **Measurement count:** Total volume of OONI measurements (pertaining to the selected country, resource, etc.)\n\nWhen trying to identify the blocking of a service (e.g. `twitter.com`), it's useful to check whether:\n\n* Measurements are annotated as `confirmed`, automatically confirming the blocking of websites\n* A large volume of measurements (in comparison to the overall measurement count) present `anomalies` (i.e. signs of potential censorship)\n\nYou can access the raw data by clicking on the bars of charts, and subsequently clicking on the relevant measurement links. \n\n# Website categories\n\n[OONI Probe](https://ooni.org/install/) users test a wide range of [websites](https://ooni.org/support/faq/#which-websites-will-i-test-for-censorship-with-ooni-probe) that fall under the following [30 standardized categories](https://github.com/citizenlab/test-lists/blob/master/lists/00-LEGEND-new_category_codes.csv).","MAT.Help.Subtitle.Categories":"Categories","MAT.CustomTooltip.ViewMeasurements":"View measurements","ReachabilityDash.Heading.CircumventionTools":"Reachability of Censorship Circumvention Tools","ReachabilityDash.CircumventionTools.Description":"The charts below display aggregate views of OONI data based on the testing of the following circumvention tools:\n\n* [Psiphon](https://ooni.org/nettest/psiphon)\n\n* [Tor](https://ooni.org/nettest/tor)\n\n* [Tor Snowflake](https://ooni.org/nettest/tor-snowflake/)\n\nPlease note that the presence of [anomalous measurements](https://ooni.org/support/faq/#what-do-you-mean-by-anomalies) is not always indicative of blocking, as [false positives](https://ooni.org/support/faq/#what-are-false-positives) can occur. Moreover, circumvention tools often have built-in circumvention techniques for evading censorship. \n\nWe therefore recommend referring to **[Tor Metrics](https://metrics.torproject.org/)** and to the **[Psiphon Data Engine](https://psix.ca/)** to view usage stats and gain a more comprehensive understanding of whether these tools work in each country.","ReachabilityDash.Form.Label.CountrySelect.AllSelected":"All countries selected","ReachabilityDash.Form.Label.CountrySelect.SearchPlaceholder":"Search","ReachabilityDash.Form.Label.CountrySelect.SelectAll":"Select All","ReachabilityDash.Form.Label.CountrySelect.SelectAllFiltered":"Select All (Filtered)","ReachabilityDash.Form.Label.CountrySelect.InputPlaceholder":"Select Countries…","ReachabilityDash.Meta.Description":"View the accessibility of censorship circumvention tools around the world through OONI data.","DateRange.Apply":"Apply","DateRange.Cancel":"Cancel","DateRange.Today":"Today","DateRange.LastWeek":"Last Week","DateRange.LastMonth":"Last Month","DateRange.LastYear":"Last Year","ThirdPartyGraph.Label.gtr":"Google (Search)","ThirdPartyGraph.Label.merit-nt":"Telescope","ThirdPartyGraph.Label.bgp":"BGP","ThirdPartyGraph.Label.ping-slash24":"Active Probing","ThirdPartyGraph.Label.cloudflare":"Cloudflare","Highlights.Political.CubaReferendum2019.Title":"2019 Constitutional Referendum","Highlights.Political.CubaReferendum2019.Text":"Blocking of independent media","Highlights.Political.VenezuelaCrisis2019.Title":"2019 Presidential Crisis","Highlights.Political.VenezuelaCrisis2019.Text":"Blocking of Wikipedia and social media","Highlights.Political.ZimbabweProtests2019.Title":"2019 Fuel Protests","Highlights.Political.ZimbabweProtests2019.Text":"Social media blocking and internet blackouts","Highlights.Political.MaliElection2018.Title":"2018 Presidential Election","Highlights.Political.MaliElection2018.Text":"Blocking of WhatsApp and Twitter","Highlights.Political.CataloniaReferendum2017.Title":"Catalonia 2017 Independence Referendum","Highlights.Political.CataloniaReferendum2017.Text":"Blocking of sites related to the referendum","Highlights.Political.IranProtests2018.Title":"2018 Anti-government Protests","Highlights.Political.IranProtests2018.Text":"Blocking of Telegram, Instagram and Tor","Highlights.Political.EthiopiaProtests2016.Title":"2016 Wave of Protests","Highlights.Political.EthiopiaProtests2016.Text":"Blocking of news websites and social media","Highlights.Political.PakistanProtests2017.Title":"2017 Protests","Highlights.Political.PakistanProtests2017.Text":"Blocking of news websites and social media","Highlights.Media.Egypt.Title":"Pervasive media censorship","Highlights.Media.Egypt.Text":"Blocking of hundreds of media websites","Highlights.Media.Venezuela.Title":"Blocking of independent media websites","Highlights.Media.Venezuela.Text":"Venezuela's economic and political crisis","Highlights.Media.SouthSudan.Title":"Blocking of foreign-based media","Highlights.Media.SouthSudan.Text":"Media accused of hostile reporting against the government","Highlights.Media.Malaysia.Title":"Blocking of media","Highlights.Media.Malaysia.Text":"1MDB scandal","Highlights.Media.Iran.Title":"Pervasive media censorship","Highlights.Media.Iran.Text":"Blocking of at least 121 news outlets","Highlights.Lgbtqi.Indonesia.Text":"Blocking of LGBTQI sites","Highlights.Lgbtqi.Iran.Text":"Blocking of Grindr","Highlights.Lgbtqi.Ethiopia.Text":"Blocking of QueerNet","Highlights.Changes.Cuba.Text":"Cuba [used to primarily serve blank block pages](https://ooni.torproject.org/post/cuba-internet-censorship-2017/), only blocking the HTTP version of websites. Now they censor access to sites that support HTTPS by means of [IP blocking](https://ooni.org/post/cuba-referendum/).","Highlights.Changes.Venezuela.Text":"Venezuelan ISPs used to primarily block sites by means of [DNS tampering](https://ooni.torproject.org/post/venezuela-internet-censorship/). Now state-owned CANTV also implements [SNI-based filtering](https://ooni.torproject.org/post/venezuela-blocking-wikipedia-and-social-media-2019/).","Highlights.Changes.Ethiopia.Text":"Ethiopia [used to block](https://ooni.org/post/ethiopia-report/) numerous news websites, LGBTQI, political opposition, and circumvention tool sites. As part of the 2018 political reforms, most of these sites have been [unblocked](https://ooni.org/post/ethiopia-unblocking/)."}}
+window.OONITranslations = {"en":{"General.OoniExplorer":"OONI Explorer","General.OK":"OK","General.Error":"Error","General.Anomaly":"Anomaly","General.Accessible":"Accessible","General.Failed":"Failed","General.Loading":"Loading…","General.NoData":"No data","General.Apply":"Apply","General.Reset":"Reset","General.Submit":"Submit","General.Close":"Close","General.Cancel":"Cancel","General.Login":"Login","General.Logout":"Logout","General.Edit":"Edit","Login.EnterEmail":"Add your email address and click the link sent to your email to log in.\n\nWe do not store email addresses.","Login.Submitted":"Your login request has been submitted. Please check your email for a link to activate and log in to your account.","Login.LoggingIn":"Logging in...","Login.Failure":"Try logging in again","Login.Success":"Successfully logged in. Redirecting back to the measurement...","SocialButtons.CTA":"Share on <facebook-link>Facebook</facebook-link> or <twitter-link>Twitter</twitter-link>","SocialButtons.Text":"Data from OONI Explorer","Tests.WebConnectivity.Name":"Web Connectivity Test","Tests.Telegram.Name":"Telegram Test","Tests.Facebook.Name":"Facebook Messenger Test","Tests.WhatsApp.Name":"WhatsApp Test","Tests.Signal.Name":"Signal Test","Tests.HTTPInvalidReqLine.Name":"HTTP Invalid Request Line Test","Tests.HTTPHeaderManipulation.Name":"HTTP Header Field Manipulation Test","Tests.NDT.Name":"NDT Speed Test","Tests.Dash.Name":"DASH Video Streaming Test","Tests.TorVanilla.Name":"Tor (Vanilla) Test","Tests.BridgeReachability.Name":"Tor Bridge Reachability Test","Tests.TCPConnect.Name":"TCP Connect Test","Tests.DNSConsistency.Name":"DNS Consistency Test","Tests.HTTPRequests.Name":"HTTP Requests Test","Tests.Psiphon.Name":"Psiphon Test","Tests.Tor.Name":"Tor Test","Tests.RiseupVPN.Name":"Riseup VPN Test","Tests.TorSnowflake.Name":"Tor Snowflake Test","Tests.DNSCheck.Name":"DNS Check","Tests.StunReachability.Name":"STUN Reachability","Tests.URLGetter.Name":"URL Getter","Tests.Groups.Webistes.Name":"Websites","Tests.Groups.Instant Messagging.Name":"Instant Messaging","Tests.Groups.Middlebox.Name":"Middleboxes","Tests.Groups.Performance.Name":"Performance","Tests.Groups.Circumvention.Name":"Circumvention","Tests.Groups.Experimental.Name":"Experimental","Tests.Groups.Legacy.Name":"Legacy","Measurement.MetaDescription":"OONI data suggests {description} on {formattedDate}, find more open data on internet censorship on OONI Explorer.","Measurement.NotFound":"Measurement not found","Measurement.Hero.Status.Confirmed":"Confirmed Blocked","Measurement.Hero.Status.Down":"Website Down","Measurement.Hero.Status.Anomaly.DNS":"DNS","Measurement.Hero.Status.Anomaly.HTTP":"HTTP","Measurement.Hero.Status.Anomaly.TCP":"TCP/IP","Measurement.CommonSummary.Label.ASN":"Network","Measurement.CommonSummary.Label.Country":"Country","Measurement.CommonSummary.Label.DateTime":"Date & Time","Measurement.DetailsHeader.Runtime":"Runtime","Measurement.Status.Hint.Websites.Censorship":"","Measurement.Status.Hint.Websites.DNS":"DNS tampering","Measurement.Status.Hint.Websites.Error":"Error in detection","Measurement.Status.Hint.Websites.HTTPdiff":"HTTP blocking (a blockpage might be served)","Measurement.Status.Hint.Websites.HTTPfail":"HTTP blocking (HTTP requests failed)","Measurement.Status.Hint.Websites.NoCensorship":"No blocking detected","Measurement.Status.Hint.Websites.TCPBlock":"TCP/IP blocking","Measurement.Status.Hint.Websites.Unavailable":"Website down","Measurement.SummaryText.Websites.Accessible":"On {date}, {WebsiteURL} was accessible when tested on {network} in {country}.","Measurement.SummaryText.Websites.Anomaly":"On {date}, {WebsiteURL} presented signs of {reason} on {network} in {country}.\n\nThis might mean that {WebsiteURL} was blocked, but <link-to-docs>false positives</link-to-docs> can occur.\n\nPlease explore the network measurement data below.","Measurement.SummaryText.Websites.Anomaly.BlockingReason.DNS":"DNS tampering","Measurement.SummaryText.Websites.Anomaly.BlockingReason.TCP":"TCP/IP blocking","Measurement.SummaryText.Websites.Anomaly.BlockingReason.HTTP-failure":"HTTP blocking (HTTP requests failed)","Measurement.SummaryText.Websites.Anomaly.BlockingReason.HTTP-diff":"HTTP blocking (a blockpage might be served)","Measurement.SummaryText.Websites.ConfirmedBlocked":"On {date}, {WebsiteURL} was blocked on {network} in {country}.\n\nThis is confirmed because a block page was served, as illustrated through the network measurement data below.","Measurement.SummaryText.Websites.Failed":"On {date}, the test for {WebsiteURL} failed on {network} in {country}.","Measurement.SummaryText.Websites.Down":"On {date}, {WebsiteURL} was down on {network} in {country}.","Measurement.Details.Websites.Failures.Heading":"Failures","Measurement.Details.Websites.Failures.Label.HTTP":"HTTP Experiment","Measurement.Details.Websites.Failures.Label.DNS":"DNS Experiment","Measurement.Details.Websites.Failures.Label.Control":"Control","Measurement.Details.Websites.Failures.Values.Null":"null","Measurement.Details.Websites.DNSQueries.Heading":"DNS Queries","Measurement.Details.Websites.DNSQueries.Label.Resolver":"Resolver","Measurement.Details.Websites.TCP.Heading":"TCP Connections","Measurement.Details.Websites.TCP.ConnectionTo":"Connection to {destination} {connectionStatus}.","Measurement.Details.Websites.TCP.ConnectionTo.Success":"succeeded","Measurement.Details.Websites.TCP.ConnectionTo.Failed":"failed","Measurement.Details.Websites.TCP.ConnectionTo.Blocked":"was blocked","Measurement.Details.Websites.HTTP.Heading":"HTTP Requests","Measurement.Details.Websites.HTTP.Label.Response":"Response","Measurement.Details.Websites.HTTP.Request.URL":"URL","Measurement.Details.Websites.HTTP.Response.Body":"Response Body","Measurement.Details.Websites.HTTP.Response.Headers":"Response Headers","Measurement.CommonDetails.Label.MsmtID":"Report ID","Measurement.CommonDetails.Label.Platform":"Platform","Measurement.CommonDetails.Label.Software":"Software Name","Measurement.CommonDetails.Label.Engine":"Measurement Engine","Measurement.CommonDetails.Value.Unavailable":"Unavailable","Measurement.CommonDetails.RawMeasurement.Heading":"Raw Measurement Data","Measurement.CommonDetails.RawMeasurement.Download":"Download JSON","Measurement.CommonDetails.RawMeasurement.Unavailable":"Unavailable","Measurement.CommonDetails.RawMeasurement.Expand":"Expand All","Measurement.CommonDetails.Label.Resolver":"Resolver","Measurement.CommonDetails.Label.ResolverASN":"Resolver ASN","Measurement.CommonDetails.Label.ResolverIP":"Resolver IP","Measurement.CommonDetails.Label.ResolverNetworkName":"Resolver Network Name","Measurement.CommonDetails.Label.UserFeedback":"User Feedback","Measurement.Hero.Status.NDT.Title":"Results","Measurement.Status.Info.Label.Download":"Download","Measurement.Status.Info.Label.Upload":"Upload","Measurement.Status.Info.Label.Ping":"Ping","Measurement.Status.Info.Label.Server":"Server","Measurement.Status.Info.NDT.Error":"Failed Test","Measurement.Details.Performance.Heading":"Performance Details","Measurement.Details.Performance.Label.AvgPing":"Average Ping","Measurement.Details.Performance.Label.MaxPing":"Max Ping","Measurement.Details.Performance.Label.MSS":"MSS","Measurement.Details.Performance.Label.RetransmitRate":"Retransmission Rate","Measurement.Details.Performance.Label.PktLoss":"Packet Loss","Measurement.Details.Performance.Label.OutOfOrder":"Out of Order","Measurement.Details.Performance.Label.Timeouts":"Timeouts","Measurement.Hero.Status.Dash.Title":"Results","Measurement.Status.Info.Label.VideoQuality":"Video Quality","Measurement.Status.Info.Label.Bitrate":"Median Bitrate","Measurement.Status.Info.Label.Delay":"Playout Delay","Measurement.Status.Hint.Telegram.Blocked":"Telegram is likely blocked","Measurement.Status.Hint.Telegram.Reachable":"Telegram is accessible","Measurement.Status.Hint.Telegram.Failed":"The Telegram test failed","Measurement.Details.SummaryText.Telegram.Reachable":"On {date}, Telegram was reachable on {network} in {country}. \n\nOONI's Telegram test successfully connected to Telegram's endpoints and web interface (web.telegram.org).","Measurement.Details.SummaryText.Telegram.AppFailure":"On {date}, the testing of Telegram's mobile app presented signs of blocking on {network} in {country}.\n\nThis might mean that Telegram's mobile app was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.Telegram.DesktopFailure":"On {date}, the testing of Telegram's web interface (web.telegram.org) presented signs of blocking on {network} in {country}.\n\nThis might mean that web.telegram.org was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.Telegram.DesktopAndAppFailure":"On {date}, the testing of Telegram's mobile app and web interface (web.telegram.org) presented signs of blocking on {network} in {country}.\n\nThis might mean that both Telegram's mobile app and web interface were blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other Telegram measurements from the same network during the same time period (if they're available).","Measurement.Details.Telegram.Endpoint.Label.Mobile":"Mobile App","Measurement.Details.Telegram.Endpoint.Label.Web":"Telegram Web","Measurement.Details.Endpoint.Status.Unknown":"Unknown","Measurement.Details.Telegram.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.Telegram.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.Telegram.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Details.Hint.WhatsApp.Reachable":"WhatsApp is accessible","Measurement.Status.Hint.WhatsApp.Blocked":"WhatsApp is likely blocked","Measurement.Status.Hint.WhatsApp.Failed":"The WhatsApp test failed","Measurement.Details.SummaryText.WhatsApp.Reachable":"On {date}, WhatsApp was reachable on {network} in {country}. \n\nOONI's WhatsApp test successfully connected to WhatsApp's endpoints, registration service and web interface (web.whatsapp.com).","Measurement.Details.SummaryText.WhatsApp.AppFailure":"On {date}, the testing of WhatsApp's mobile app presented signs of blocking on {network} in {country}.\n\nThis might mean that WhatsApp's mobile app was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.WhatsApp.DesktopFailure":"On {date}, the testing of WhatsApp's web interface (web.whatsapp.com) presented signs of blocking on {network} in {country}.\n\nThis might mean that web.whatsapp.com was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.WhatsApp.DesktopAndAppFailure":"On {date}, the testing of WhatsApp's mobile app and web interface (web.whatsapp.com) presented signs of blocking on {network} in {country}.\n\nThis might mean that both WhatsApp's mobile app and web interface were blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.\n\nPlease explore the network measurement data below. Check other WhatsApp measurements from the same network during the same time period (if they're available).","Measurement.Details.WhatsApp.Endpoint.Label.Mobile":"Mobile App","Measurement.Details.WhatsApp.Endpoint.Label.Web":"WhatsApp Web","Measurement.Details.WhatsApp.Endpoint.Label.Registration":"Registration","Measurement.Details.WhatsApp.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.WhatsApp.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.WhatsApp.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Status.Hint.FacebookMessenger.Reachable":"Facebook Messenger is accessible","Measurement.Status.Hint.FacebookMessenger.Blocked":"Facebook Messenger is likely blocked","Measurement.Status.Hint.FacebookMessenger.Failed":"The Facebook Messenger test failed","Measurement.Details.SummaryText.FacebookMessenger.Reachable":"On {date}, Facebook Messenger was reachable on {network} in {country}.","Measurement.Details.SummaryText.FacebookMessenger.TCPFailure":"TCP connections to Facebook's endpoints failed.","Measurement.Details.SummaryText.FacebookMessenger.DNSFailure":"DNS lookups did not resolve to Facebook IP addresses.","Measurement.Details.SummaryText.FacebookMessenger.DNSSuccess":"DNS lookups resolved to Facebook IP addresses.","Measurement.Details.SummaryText.FacebookMessenger.TCPSuccess":"TCP connections to Facebook's enpoints succeeded.","Measurement.Details.FacebookMessenger.TCP.Label.Title":"TCP connections","Measurement.Details.FacebookMessenger.DNS.Label.Title":"DNS lookups","Measurement.Details.FacebookMessenger.TCPFailed":"TCP connections failed","Measurement.Details.FacebookMessenger.DNSFailed":"DNS lookups failed","Measurement.Details.FacebookMessenger.Endpoint.Status.Heading":"Endpoint Status","Measurement.Details.FacebookMessenger.Endpoint.ConnectionTo.Failed":"Connection to {destination} failed.","Measurement.Details.FacebookMessenger.Endpoint.ConnectionTo.Successful":"Connection to {destination} was successful.","Measurement.Status.Hint.Signal.Blocked":"Signal is likely blocked","Measurement.Status.Hint.Signal.Reachable":"Signal is accessible","Measurement.Status.Hint.Signal.Failed":"The Signal test failed","Measurement.Details.SummaryText.Signal.Reachable":"On {date}, [Signal](https://signal.org/) was reachable on {network} in {country}. \n\nThe [OONI Probe Signal test](https://ooni.org/nettest/signal) successfully connected to Signal's endpoints.","Measurement.Details.SummaryText.Signal.Blocked":"On {date}, the testing of the [Signal app](https://signal.org/) presented signs of blocking on {network} in {country}.\n\nThis might mean that Signal was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other Signal measurements from the same network during the same time period (if they're available).","Measurement.Hero.Status.HTTPHeaderManipulation.NoMiddleBoxes":"No middleboxes detected","Measurement.HTTPHeaderManipulation.NoMiddleBoxes.SummaryText":"On {date}, no network anomaly was detected on {network} in {country} when communicating with our servers.","Measurement.Hero.Status.HTTPHeaderManipulation.MiddleboxesDetected":"Network tampering","Measurement.HTTPHeaderManipulation.MiddleBoxesDetected.SummaryText":"On {date}, network traffic was manipulated when contacting our control servers. \n\nThis means that there might be a middlebox on {network} in {country}, which could be responsible for censorship and/or surveillance.","Measurement.Hero.Status.HTTPInvalidReqLine.NoMiddleBoxes":"No middleboxes detected","Measurement.HTTPInvalidReqLine.NoMiddleBoxes.SummaryText":"On {date}, no network anomaly was detected on {network} in {country} when communicating with our servers.","Measurement.Hero.Status.HTTPInvalidReqLine.MiddleboxesDetected":"Network tampering","Measurement.HTTPInvalidReqLine.MiddleboxesDetected.SummaryText":"On {date}, network traffic was manipulated when contacting our control servers. \n\nThis means that there might be a middlebox on {network} in {country}, which could be responsible for censorship and/or surveillance.","Measurement.HTTPInvalidReqLine.YouSent":"You Sent","Measurement.HTTPInvalidReqLine.YouReceived":"You Received","Measurement.Hero.Status.TorVanilla.Blocked":"Tor is likely blocked","Measurement.Hero.Status.TorVanilla.Reachable":"Tor is accessible","Measurement.Details.SummaryText.TorVanilla.Blocked":"On {date}, OONI's Vanilla Tor test did not manage to bootstrap a connection to the [Tor network](https://www.torproject.org/).\n\nThis might mean that access to the Tor network was blocked on {network} in {country}, but [false positives can occur](https://ooni.org/support/faq/#why-do-false-positives-occur).\n\nPlease explore the network measurement data below. Check other Tor measurements from the same network during the same time period (if they're available).","Measurement.Details.SummaryText.TorVanilla.Reachable":"OONI's Vanilla Tor test successfully bootstrapped a connection to the [Tor network](https://www.torproject.org/).\n\nThis means that the Tor network was reachable from {network} in {country} on {date}.","Measurement.Details.VanillaTor.Endpoint.Label.Reachability":"Reachability","Measurement.Status.Hint.Psiphon.Reachable":"Psiphon works","Measurement.Status.Hint.Psiphon.Blocked":"Psiphon is likely blocked","Measurement.Status.Hint.Psiphon.BootstrappingError":"Psiphon is likely blocked (bootstrap error)","Measurement.Details.SummaryText.Psiphon.OK":"On {date}, [Psiphon](https://psiphon.ca/) worked on {network} in {country}.\n\nThe [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was able to successfully bootstrap Psiphon and ensure that the app works.","Measurement.Details.SummaryText.Psiphon.Blocked":"On {date}, [Psiphon](https://psiphon.ca/) did not appear to work on {network} in {country}.\n\nWhile the [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was able to bootstrap Psiphon, it was unable to fetch a webpage from the internet. \n\nThis suggests that the Psiphon app may have been blocked on this network. \n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. Please explore the network measurement data below and compare it with other relevant measurements (if they're available).","Measurement.Details.SummaryText.Psiphon.BootstrappingError":"On {date}, [Psiphon](https://psiphon.ca/) did not work on {network} in {country}.\n\nThe [OONI Probe Psiphon test](https://ooni.org/nettest/psiphon/) was unable to bootstrap Psiphon.\n\nThis suggests that the Psiphon app may have been blocked on this network.\n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. Please explore the network measurement data below and compare it with other relevant measurements (if they're available).","Measurement.Details.Psiphon.BootstrapTime.Label":"Bootstrap Time","Measurement.Status.Hint.Tor.Reachable":"Tor works","Measurement.Status.Hint.Tor.Blocked":"Tor is likely blocked","Measurement.Status.Hint.Tor.Error":"Tor test failed","Measurement.Details.SummaryText.Tor.OK":"On {date}, [Tor](https://www.torproject.org/) worked on {network} in {country}.\n\nAs part of [OONI Probe Tor testing](https://ooni.org/nettest/tor/), all reachability measurements of selected Tor directory authorities and bridges were successful.","Measurement.Details.SummaryText.Tor.Blocked":"On {date}, [Tor](https://www.torproject.org/) did not appear to work on {network} in {country}.\n\nThe [OONI Probe Tor test](https://ooni.org/nettest/tor/) failed in performing certain measurements. More details are available through the network measurement data provided below.\n\nThis suggests that Tor may have been blocked on this network.\n\nHowever, [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur.","Measurement.Details.SummaryText.Tor.Error":"On {date}, the Tor test failed on {network} in {country}.","Measurement.Details.Tor.Bridges.Label.Title":"Tor Browser Bridges","Measurement.Details.Tor.Bridges.Label.OK":"{bridgesAccessible}/{bridgesTotal} OK","Measurement.Details.Tor.DirAuth.Label.Title":"Tor Directory Authorities","Measurement.Details.Tor.DirAuth.Label.OK":"{dirAuthAccessible}/{dirAuthTotal} OK","Measurement.Details.Tor.Table.Header.Name":"Name","Measurement.Details.Tor.Table.Header.Address":"Address","Measurement.Details.Tor.Table.Header.Type":"Type","Measurement.Status.Hint.TorSnowflake.Reachable":"Tor Snowflake works","Measurement.Status.Hint.TorSnowflake.Blocked":"Tor Snowflake does not work","Measurement.Status.Hint.TorSnowflake.Error":"Tor Snowflake test failed","Measurement.Details.SummaryText.TorSnowflake.OK":"On {date}, [Tor Snowflake](https://www.torproject.org/) worked on {network} in {country}.\n\nThe [OONI Probe Tor Snowflake test](https://ooni.org/nettest/torsf/) was able to successfully bootstrap Snowflake.","Measurement.Details.SummaryText.TorSnowflake.Blocked":"On {date}, [Tor Snowflake](https://www.torproject.org/) does not work on {network} in {country}.\n\nThe [OONI Probe Tor Snowflake test](https://ooni.org/nettest/torsf/) failed to bootstrap Snowflake.","Measurement.Details.SummaryText.TorSnowflake.Error":"On {date}, the Tor Snowflake test failed on {network} in {country}.","Measurement.Details.TorSnowflake.BootstrapTime.Label":"Bootstrap Time","Measurement.Details.TorSnowflake.Error.Label":"Failure","Measurement.Metadata.TorSnowflake.Reachable":"Tor Snowflake was reachable in {country}","Measurement.Metadata.TorSnowflake.UnReachable":"Tor Snowflake was NOT reachable in {country}","Measurement.Metadata.TorSnowflake.Error":"Tor Snowflake test failed in {country}","Measurement.Status.Hint.RiseupVPN.Reachable":"RiseupVPN works","Measurement.Status.Hint.RiseupVPN.Blocked":"RiseupVPN is likely blocked","Measurement.Status.Hint.RiseupVPN.Failed":"The RiseupVPN test failed","Measurement.Details.SummaryText.RiseupVPN.OK":"On {date}, [RiseupVPN](https://riseup.net/vpn) was reachable on {network} in {country}. \n\nThe [OONI Probe RiseupVPN test](https://ooni.org/nettest/riseupvpn/) successfully connected to RiseupVPN's bootstrap servers and gateways.","Measurement.Details.SummaryText.RiseupVPN.Blocked":"On {date}, the testing of [RiseupVPN](https://riseup.net/vpn) presented signs of blocking on {network} in {country}.\n\nThis might mean that RiseupVPN was blocked, but [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur) can occur. \n\nPlease explore the network measurement data below. Check other RiseupVPN measurements from the same network during the same time period (if they're available).","Measurement.Metadata.RiseupVPN.Reachable":"RiseupVPN was reachable in {country}","Measurement.Metadata.RiseupVPN.Blocked":"RiseupVPN was not reachable in {country}","Measurement.Hero.Status.Default":"Measurement Report","Measurement.Feedback.Title":"Verify the measurement","Measurement.Feedback.Description":"Please share feedback based on what you see in the measurement data.","Measurement.Feedback.ok":"It's OK","Measurement.Feedback.down":"It's down","Measurement.Feedback.down.unreachable":"It's unreachable","Measurement.Feedback.down.misconfigured":"It's misconfigured","Measurement.Feedback.blocked":"It's blocked","Measurement.Feedback.blocked.tcp":"TCP/IP blocking","Measurement.Feedback.blocked.tls":"TLS blocking","Measurement.Feedback.blocked.blockpage":"Block page","Measurement.Feedback.blocked.blockpage.http":"HTTP block page","Measurement.Feedback.blocked.blockpage.dns":"DNS block page","Measurement.Feedback.blocked.blockpage.server_side":"Server-side block page","Measurement.Feedback.blocked.blockpage.server_side.captcha":"CAPTCHA","Measurement.Feedback.blocked.dns":"DNS blocking without block page","Measurement.Feedback.blocked.dns.inconsistent":"Inconsistent DNS response","Measurement.Feedback.blocked.dns.nxdomain":"NXDOMAIN error","Measurement.Feedback.Success.Title":"Thank you!","Measurement.Feedback.Success.Description":"Your feedback will help improve OONI measurements.","Measurement.Feedback.Failure":"Something went wrong, please try again.","Measurement.Feedback.Login.Title":"Please log in to continue","Measurement.Feedback.Login.Confirmation.Title":"Login link sent","Measurement.Feedback.Login.Confirmation.Text":"Please check your email for a link to log in to your account.","Measurement.Feedback.Login.Description":"We will send a link to your email. We do not store emails.","Measurement.Feedback.ExistingFeedback":"Your previous feedback","Navbar.Search":"Search","Navbar.Countries":"Countries","Navbar.Charts.Circumvention":"Circumvention Charts","Navbar.Charts.MAT":"MAT Charts","Network.Summary.TotalMeasurements":"Total number of measurements: **{measurementsTotal}**","Network.Summary.FirstMeasurement":"Date of the first measurement: **{formattedDate}**","Network.Summary.Countries":"Network observed in countries:","Network.Summary.Country.Measurements":"({measurementsTotal} measurements)","Network.NoData.Title":"Let's collect more data!","Network.NoData.Text":"We don't have enough data for this network to show the charts. Please run OONI Probe to collect more measurements.","Footer.Text.Slogan":"Global community measuring internet censorship around the world.","Footer.Heading.About":"About","Footer.Heading.OONIProbe":"OONI Probe","Footer.Heading.Updates":"Updates","Footer.Link.About":"OONI","Footer.Link.DataPolicy":"Data Policy","Footer.Link.DataLicense":"Data License","Footer.Link.Contact":"Contact","Footer.Link.Probe":"Install","Footer.Link.Tests":"Tests","Footer.Link.Code":"Source code","Footer.Link.API":"API","Footer.Link.Blog":"Blog","Footer.Link.Twitter":"Twitter","Footer.Link.MailingList":"Mailing list","Footer.Link.Slack":"Slack","Footer.Text.Copyright":"© {currentYear} Open Observatory of Network Interference (OONI)","Footer.Text.CCommons":"Content available under a Creative Commons license.","Footer.Text.Version":"Version","CategoryCode.ALDR.Name":"Alcohol & Drugs","CategoryCode.REL.Name":"Religion","CategoryCode.PORN.Name":"Pornography","CategoryCode.PROV.Name":"Provocative Attire","CategoryCode.POLR.Name":"Political Criticism","CategoryCode.HUMR.Name":"Human Rights Issues","CategoryCode.ENV.Name":"Environment","CategoryCode.MILX.Name":"Terrorism and Militants","CategoryCode.HATE.Name":"Hate Speech","CategoryCode.NEWS.Name":"News Media","CategoryCode.XED.Name":"Sex Education","CategoryCode.PUBH.Name":"Public Health","CategoryCode.GMB.Name":"Gambling","CategoryCode.ANON.Name":"Anonymization and circumvention tools","CategoryCode.DATE.Name":"Online Dating","CategoryCode.GRP.Name":"Social Networking","CategoryCode.LGBT.Name":"LGBTQ+","CategoryCode.FILE.Name":"File-sharing","CategoryCode.HACK.Name":"Hacking Tools","CategoryCode.COMT.Name":"Communication Tools","CategoryCode.MMED.Name":"Media sharing","CategoryCode.HOST.Name":"Hosting and Blogging Platforms","CategoryCode.SRCH.Name":"Search Engines","CategoryCode.GAME.Name":"Gaming","CategoryCode.CULTR.Name":"Culture","CategoryCode.ECON.Name":"Economics","CategoryCode.GOVT.Name":"Government","CategoryCode.COMM.Name":"E-commerce","CategoryCode.CTRL.Name":"Control content","CategoryCode.IGO.Name":"Intergovernmental Organizations","CategoryCode.MISC.Name":"Miscellaneous content","CategoryCode.ALDR.Description":"Sites devoted to the use, paraphernalia, and sale of drugs and alcohol irrespective of the local legality.","CategoryCode.REL.Description":"Sites devoted to discussion of religious issues, both supportive and critical, as well as discussion of minority religious groups.","CategoryCode.PORN.Description":"Hard-core and soft-core pornography.","CategoryCode.PROV.Description":"Websites which show provocative attire and portray women in a sexual manner, wearing minimal clothing.","CategoryCode.POLR.Description":"Content that offers critical political viewpoints. Includes critical authors and bloggers, as well as oppositional political organizations. Includes pro-democracy content, anti-corruption content as well as content calling for changes in leadership, governance issues, legal reform, etc.","CategoryCode.HUMR.Description":"Sites dedicated to discussing human rights issues in various forms. Includes women's rights and rights of minority ethnic groups.","CategoryCode.ENV.Description":"Pollution, international environmental treaties, deforestation, environmental justice, disasters, etc.","CategoryCode.MILX.Description":"Sites promoting terrorism, violent militant or separatist movements.","CategoryCode.HATE.Description":"Content that disparages particular groups or persons based on race, sex, sexuality or other characteristics.","CategoryCode.NEWS.Description":"This category includes major news outlets (BBC, CNN, etc.) as well as regional news outlets and independent media.","CategoryCode.XED.Description":"Includes contraception, abstinence, STDs, healthy sexuality, teen pregnancy, rape prevention, abortion, sexual rights, and sexual health services.","CategoryCode.PUBH.Description":"HIV, SARS, bird flu, centers for disease control, World Health Organization, etc.","CategoryCode.GMB.Description":"Online gambling sites. Includes casino games, sports betting, etc.","CategoryCode.ANON.Description":"Sites that provide tools used for anonymization, circumvention, proxy-services and encryption.","CategoryCode.DATE.Description":"Online dating services which can be used to meet people, post profiles, chat, etc.","CategoryCode.GRP.Description":"Social networking tools and platforms.","CategoryCode.LGBT.Description":"A range of gay-lesbian-bisexual-transgender queer issues. (Excluding pornography)","CategoryCode.FILE.Description":"Sites and tools used to share files, including cloud-based file storage, torrents and P2P file-sharing tools.","CategoryCode.HACK.Description":"Sites dedicated to computer security, including news and tools. Includes malicious and non-malicious content.","CategoryCode.COMT.Description":"Sites and tools for individual and group communications. Includes webmail, VoIP, instant messaging, chat and mobile messaging applications.","CategoryCode.MMED.Description":"Video, audio or photo sharing platforms.","CategoryCode.HOST.Description":"Web hosting services, blogging and other online publishing platforms.","CategoryCode.SRCH.Description":"Search engines and portals.","CategoryCode.GAME.Description":"Online games and gaming platforms, excluding gambling sites.","CategoryCode.CULTR.Description":"Content relating to entertainment, history, literature, music, film, books, satire and humour.","CategoryCode.ECON.Description":"General economic development and poverty related topics, agencies and funding opportunities.","CategoryCode.GOVT.Description":"Government-run websites, including military sites.","CategoryCode.COMM.Description":"Websites of commercial services and products.","CategoryCode.CTRL.Description":"Benign or innocuous content used as a control.","CategoryCode.IGO.Description":"Websites of intergovernmental organizations such as the United Nations.","CategoryCode.MISC.Description":"Sites that don't fit in any category. (XXX Things in here should be categorised)","Country.Heading.Overview":"Overview","Country.Heading.Websites":"Websites","Country.Heading.Apps":"Apps","Country.Heading.Shutdowns":"Shutdowns","Country.Heading.NetworkProperties":"Networks","Country.Overview.Heading.NwInterference":"In a nutshell","Country.Overview.NwInterference.Middleboxes.Blocked":"Middleboxes were detected on {middleboxCount} network(s)","Country.Overview.NwInterference.Middleboxes.Normal":"No middleboxes were detected on tested networks","Country.Overview.NwInterference.Middleboxes.NoData":"Not enough data available on middleboxes","Country.Overview.NwInterference.IM.Blocked":"Instant messaging apps were likely blocked","Country.Overview.NwInterference.IM.Normal":"No instant messaging apps were blocked on tested networks","Country.Overview.NwInterference.IM.NoData":"Not enough data available on instant messaging apps","Country.Overview.NwInterference.CircumventionTools.Blocked":"Circumvention tools were likely blocked","Country.Overview.NwInterference.CircumventionTools.Normal":"No circumvention tools were blocked on tested networks","Country.Overview.NwInterference.CircumventionTools.NoData":"Not enough data available on circumvention tools","Country.Overview.NwInterference.Websites.Blocked":"OONI data confirms the blocking of websites","Country.Overview.NwInterference.Websites.Normal":"The blocking of websites is not confirmed","Country.Overview.NwInterference.Websites.NoData":"Not enough data available on blocked websites","Country.Overview.Heading.TestsByClass":"Measurement Coverage","Country.Overview.Heading.TestsByClass.Description":"The graph below provides an overview of OONI Probe measurement coverage. It shows how many results have been collected from each OONI Probe test category, as well as how many networks have been covered by tests. \n\nBy looking at this graph, you can understand if there is enough data to draw meaningful conclusions. If there is not enough data and you are in the country in question, [install OONI Probe](https://ooni.org/install), run tests, and contribute data!","Country.Overview.TestsByClass.Websites":"Websites","Country.Overview.TestsByClass.InstantMessaging":"Instant Messaging","Country.Overview.TestsByClass.Performance":"Performance","Country.Overview.TestsByClass.Middleboxes":"Middleboxes","Country.Overview.TestsByClass.Circumvention":"Circumvention Tools","Country.Overview.FeaturedResearch":"Research Reports","Country.Overview.FeaturedResearch.None":"We haven't published a research report based on OONI data from this country yet. \n\nWe encourage you to use OONI data for your research!","Country.Overview.SummaryTextTemplate":"OONI Probe users in **{countryName}** have collected [**{measurementCount}** measurements]({linkToMeasurements}) from **{networkCovered}** local networks.\n\nExplore the data below to check the accessibility and/or blocking of sites and services.","Country.Overview.NoData.Title":"Let's collect more data!","Country.Overview.NoData.CallToAction":"We don’t have enough measurements for **{country}** to show a chart. If you are in {country} or know people there, tell them to run OONI Probe to collect more measurements.","Country.Overview.NoData.Button.InstallProbe":"Install OONI Probe","Country.Overview.NoData.Button.OoniRunLink":"Create OONI Run Link","Country.Meta.Title":"Internet Censorship in {countryName} - OONI Explorer","Country.Meta.Description":"OONI Probe users in {countryName} have collected {measurementCount} measurements from {networkCount} local networks. Explore the data on OONI Explorer.","Country.PeriodFilter.Label":"Show results from","Country.PeriodFilter.Option.30Days":"Last 30 Days","Country.PeriodFilter.Option.2Months":"Last 2 Months","Country.PeriodFilter.Option.3Months":"Last 3 Months","Country.PeriodFilter.Option.6Months":"Last 6 Months","Country.Websites.Description":"Check whether websites have been blocked.\n\nTesting methodology: OONI's [Web Connectivity test](https://ooni.org/nettest/web-connectivity/), designed to measure the DNS, HTTP, and TCP/IP blocking of websites. \n\nTested websites: [Citizen Lab test lists](https://github.com/citizenlab/test-lists)\n\nIf you'd like to see results on the testing of different websites, please [contribute to test lists](https://ooni.org/get-involved/contribute-test-lists/) or test the sites of your choice via the [OONI Probe mobile app](https://ooni.org/install/). \n\nPlease note that unless a block page is served, some anomalous measurements may contain [false positives](https://ooni.org/support/faq/#why-do-false-positives-occur). We therefore encourage you to examine anomalous measurements in depth and over time.","Country.Websites.Description.MoreLinkText":"","Country.Websites.Heading.BlockedByCategory":"Categories of Blocked Websites","Country.Websites.BlockedByCategory.Description":"Websites that fall under the following categories were blocked on the {selectedASN} network.","Country.Websites.TestedWebsitesCount":"URLs tested","Country.Websites.Labels.ResultsPerPage":"Results per page","Country.Websites.URLSearch.Placeholder":"Search for URL","Country.Websites.URLCharts.Legend.Label.Blocked":"Confirmed Blocked","Country.Websites.URLCharts.ExploreMoreMeasurements":"Explore more measurements","Country.Websites.URLCharts.Pagination.Previous":"Previous Page","Country.Websites.URLCharts.Pagination.Next":"Next Page","Country.Apps.Description":"Check whether instant messaging apps and circumvention tools are blocked.\n\nThe following results were collected through the use of [OONI Probe tests](https://ooni.org/nettest/) designed to measure the blocking of WhatsApp, Facebook Messenger, Telegram and Signal. \n\nWe also share results on the testing of circumvention tools, like [Tor](https://www.torproject.org/) and [Psiphon](https://psiphon.ca/).","Country.Apps.Label.LastTested":"Last tested","Country.Apps.Label.TestedNetworks":"tested networks","Country.Apps.Button.ShowMore":"Show More","Country.Shutdowns":"Internet shutdowns","Country.Shutdowns.Description":"Monitor Internet shutdowns through public, third-party data sources. \n\nThe following graph shares data from Georgia Tech's [Internet Outage Detection and Analysis (IODA)](https://ioda.inetintel.cc.gatech.edu/) project, [Google Transparency Reports (Google traffic data)](https://transparencyreport.google.com/traffic/overview?hl=en), and [Cloudflare Radar](https://radar.cloudflare.com/).\n\nIf you notice a drop in all signals, that might serve as an indicator of an Internet shutdown. \n\nLearn more about Internet shutdowns and their impact through the [#KeepItOn campaign](https://www.accessnow.org/keepiton/).","Country.NetworkProperties.Description":"Check the speed and performance of networks.\n\nThe following results were collected through the use of [OONI Probe's performance and middlebox tests](https://ooni.org/nettest/). You can check the speed and performance of tested networks, as well as video streaming performance. \n\nYou can also learn whether middleboxes were detected in tested networks. Middleboxes are network appliances that can be used for a variety of networking purposes (such as caching), but sometimes they're used to implement internet censorship and/or surveillance.","Country.NetworkProperties.Heading.Summary":"Summary","Country.NetworkProperties.Heading.Networks":"Networks","Country.NetworkProperties.InfoBox.Label.AverageDownload":"Average Download","Country.NetworkProperties.InfoBox.Label.AverageUpload":"Average Upload","Country.NetworkProperties.InfoBox.Label.Covered":"Covered","Country.NetworkProperties.InfoBox.Label.Middleboxes":"Middleboxes detected","Country.NetworkProperties.InfoBox.Units.Mbits":"Mbit/s","Country.NetworkProperties.InfoBox.Units.Networks.Singular":"Network","Country.NetworkProperties.InfoBox.Units.Networks.Plural":"Networks","Country.NetworkProperties.InfoBox.Label.AverageStreaming":"Average Streaming","Country.NetworkProperties.InfoBox.Label.AveragePing":"Average Ping","Country.NetworkProperties.InfoBox.Units.Milliseconds":"ms","Country.NetworkProperties.InfoBox.Label.Middleboxes.Found":"Middleboxes detected","Country.NetworkProperties.InfoBox.Label.Middleboxes.NotFound":"No middleboxes detected","Country.NetworkProperties.Button.ShowMore":"Show more networks","Country.Label.NoData":"No Data Available","Search.PageTitle":"Search through millions of Internet censorship measurements","Search.Sidebar.Domain":"Domain","Search.Sidebar.Domain.Placeholder":"e.g. twitter.com or 1.1.1.1","Search.Sidebar.Domain.Error":"Please enter a valid domain name or IP address, such as twitter.com or 1.1.1.1","Search.Sidebar.Input":"Input","Search.Sidebar.Input.Placeholder":"e.g., https://twitter.com/","Search.Sidebar.Input.Error":"Please enter a full URL (e.g. https://twitter.com/) with a trailing slash (`/`) or an IP address","Search.Sidebar.Categories":"Website Categories","Search.Sidebar.Categories.All":"Any","Search.Sidebar.Status":"Status","Search.Sidebar.TestName":"Test Name","Search.Sidebar.TestName.AllTests":"Any","Search.Sidebar.Country":"Country","Search.Sidebar.Country.AllCountries":"Any","Search.Sidebar.ASN":"ASN","Search.Sidebar.ASN.example":"e.g. AS30722","Search.Sidebar.ASN.Error":"Valid formats: AS1234, 1234","Search.Sidebar.From":"From","Search.Sidebar.Until":"Until","Search.Sidebar.HideFailed":"Hide failed measurements","Search.Sidebar.Button.FilterResults":"Filter Results","Search.FilterButton.AllResults":"All Results","Search.FilterButton.Confirmed":"Confirmed","Search.FilterButton.Anomalies":"Anomalies","Search.FilterButton.Search":"Search","Search.Filter.SortBy":"Sort by","Search.Filter.SortBy.Date":"Date","Search.WebConnectivity.Results.Blocked":"Confirmed","Search.HTTPRequests.Results.Anomaly":"","Search.HTTPRequests.Results.Blocked":"","Search.HTTPRequests.Results.Error":"","Search.HTTPRequests.Results.Reachable":"","Search.NDT.Results":"","Search.DASH.Results":"","Search.VanillaTor.Results":"","Search.BridgeReachability.Results":"","Search.LegacyTests.Results":"","Search.Results.Empty.Heading":"No Results Found","Search.Results.Empty.Description":"Please try changing the filters to get better results.","Search.Button.LoadMore":"Load more","Search.Error.Message":"This query took too long to complete. Please try adjusting the search filters or view the example queries in the [Highlights section of the homepage](/#highlights).\n\nIf you are interested in using OONI data in batch, we recommend the [ooni-data Amazon S3 bucket](https://ooni.org/post/mining-ooni-data/) or the [aggregation API](https://api.ooni.io/apidocs/#/default/get_api_v1_aggregation).\n\nWe are working on improving the performance of OONI Explorer. To track our work on this, [see the open issues on the ooni/api repository](https://github.com/ooni/api/issues?q=is%3Aissue+is%3Aopen+label%3Aoptimization).","Search.Error.Details.Label":"Server Response","Home.Banner.Title.UncoverEvidence":"Uncover evidence of internet censorship worldwide","Home.Banner.Subtitle.ExploreCensorshipEvents":"Open data collected by the global OONI community","Home.Banner.Button.Explore":"Explore","Home.Banner.Stats.Measurements":"Measurements","Home.Banner.Stats.Countries":"Countries","Home.Banner.Stats.Networks":"Networks","Home.About.SummaryText":"OONI Explorer is an open data resource on internet censorship around the world. \n\nSince 2012, millions of network measurements have been collected from more than 200 countries. OONI Explorer sheds light on internet censorship and other forms of network interference worldwide.\n\nTo contribute to this open dataset, [install OONI Probe](https://ooni.org/install/) and run tests!","Home.Websites&Apps.Title":"Blocking of Websites & Apps","Home.Websites&Apps.SummaryText":"Discover blocked websites around the world. Check whether WhatsApp, Facebook Messenger, and Telegram are blocked.","Home.Search&Filter.Title":"Search","Home.Search&Filter.SummaryText":"Explore OONI measurements with a powerful search tool. View the most recently blocked websites. Compare internet censorship across networks.","Home.NetworkProperties.Title":"Network Performance","Home.NetworkProperties.SummaryText":"Check the speed and performance of thousands of networks around the world. Explore data on video streaming performance.","Home.MonthlyStats.Title":"Monthly coverage worldwide","Home.MonthlyStats.SummaryText":"OONI Explorer hosts millions of network measurements collected from more than 200 countries since 2012. Every day, OONI Explorer is updated with new measurements!\n\nLast month, {measurementCount} OONI Probe measurements were collected from {networkCount} networks in {countryCount} countries. Explore the monthly usage of [OONI Probe](https://ooni.org/install/) through the stats below.","Home.Highlights.CTA":"We encourage you to <link-to-search>explore OONI measurements</link-to-search> to find more highlights!","Home.Highlights.Title":"Highlights","Home.Highlights.Description":"What can you learn from OONI Explorer? \n\nBelow we share some stories from [research reports](https://ooni.org/post/) based on OONI data.\n\nWe share these case studies to demonstrate how OONI's openly available data can be used and what types of stories can be told. \n\nWe encourage you to explore OONI data, discover more censorship cases, and to use OONI data as part of your research and/or advocacy.","Home.Highlights.Political":"Censorship during political events","Home.Highlights.Political.Description":"Internet censorship sometimes occurs in response to or in anticipation of political events, such as elections, protests, and riots. Below we share a few cases detected via OONI data and correlated with political events.","Home.Highlights.Media":"Media censorship","Home.Highlights.Media.Description":"Press freedom is threatened in countries that experience the blocking of media websites. Below we share a few cases detected through OONI data.","Home.Highlights.LGBTQI":"Blocking of LGBTQI sites","Home.Highlights.LGBTQI.Description":"Minority group sites are blocked around the world. Below we share a few cases on the blocking of LGBTQI sites.","Home.Highlights.Changes":"Censorship changes","Home.Highlights.Changes.Description":"OONI measurements have been collected on a continuous basis since 2012, enabling the identification of censorship changes around the world. Some examples include:","Home.Meta.Description":"OONI Explorer is an open data resource on Internet censorship around the world consisting of millions of measurements on network interference.","Home.Highlights.Explore":"Explore","Home.Highlights.ReadReport":"Read report","Countries.Heading.JumpToContinent":"Jump to continent","Countries.Search.NoCountriesFound":"No countries found with {searchTerm}","Countries.Search.Placeholder":"Search for countries","Countries.PageTitle":"Internet Censorship around the world","Error.404.PageNotFound":"Page Not Found","Error.404.GoBack":"Go back","Error.404.Heading":"The requested page does not exist","Error.404.Message":"We could not find the content you were looking for. Maybe try {measurmentLink} or look at {homePageLink}.","Error.404.MeasurmentLinkText":"exploring some measurement","Error.404.HomepageLinkText":"the homepage","MAT.Title":"OONI Measurement Aggregation Toolkit (MAT)","MAT.SubTitle":"Create charts based on aggregate views of real-time OONI data from around the world","MAT.JSONData":"JSON Data","MAT.CSVData":"CSV Data","MAT.Form.Label.XAxis":"Columns","MAT.Form.Label.YAxis":"Rows","MAT.Form.Label.TimeGrain":"Time Granularity","MAT.Form.Label.AxisOption.domain":"Domain","MAT.Form.Label.AxisOption.input":"Input","MAT.Form.Label.AxisOption.measurement_start_day":"Measurement Day","MAT.Form.Label.AxisOption.probe_cc":"Countries","MAT.Form.Label.AxisOption.category_code":"Website Categories","MAT.Form.Label.AxisOption.probe_asn":"ASN","MAT.Form.TimeGrainOption.hour":"Hour","MAT.Form.TimeGrainOption.day":"Day","MAT.Form.TimeGrainOption.week":"Week","MAT.Form.TimeGrainOption.month":"Month","MAT.Form.ConfirmationModal.Title":"Are you sure?","MAT.Form.ConfirmationModal.Message":"Duration too long. This can potentially slow down the page","MAT.Form.ConfirmationModal.No":"No","MAT.Form.ConfirmationModal.Button.Yes":"Yes","MAT.Form.Submit":"Show Chart","MAT.Form.All":"All","MAT.Form.AllCountries":"All Countries","MAT.Table.Header.ok_count":"OK Count","MAT.Table.Header.anomaly_count":"Anomaly Count","MAT.Table.Header.confirmed_count":"Confirmed Count","MAT.Table.Header.failure_count":"Failure Count","MAT.Table.Header.measurement_count":"Measurement Count","MAT.Table.Header.input":"URL","MAT.Table.Header.category_code":"Category Code","MAT.Table.Header.probe_cc":"Country","MAT.Table.Header.probe_asn":"ASN","MAT.Table.Header.blocking_type":"Blocking Type","MAT.Table.Header.domain":"Domain","MAT.Table.FilterPlaceholder":"Search {count} records…","MAT.Table.Search":"Search:","MAT.Table.Filters":"Filters","MAT.Charts.NoData.Title":"No Data Found","MAT.Charts.NoData.Description":"We are not able to produce a chart based on the selected filters. Please change the filters and try again.","MAT.Charts.NoData.Details":"Details:","MAT.Help.Box.Title":"Help","MAT.Help.Title":"FAQs","MAT.Help.Content":"# What is the MAT?\n\nOONI's Measurement Aggregation Toolkit (MAT) is a tool that enables you to generate your own custom charts based on **aggregate views of real-time OONI data** collected from around the world.\n\nOONI data consists of network measurements collected by [OONI Probe](https://ooni.org/install/) users around the world. \n\nThese measurements contain information about various types of **internet censorship**, such as the [blocking of websites and apps](https://ooni.org/nettest/) around the world. \n\n# Who is the MAT for?\n\nThe MAT was built for researchers, journalists, and human rights defenders interested in examining internet censorship around the world.\n\n# Why use the MAT?\n\nWhen examining cases of internet censorship, it's important to **look at many measurements at once** (\"in aggregate\") in order to answer key questions like the following:\n\n* Does the testing of a service (e.g. Facebook) present **signs of blocking every time that it is tested** in a country? This can be helpful for ruling out [false positives](https://ooni.org/support/faq/#what-are-false-positives).\n* What types of websites (e.g. human rights websites) are blocked in each country?\n* In which countries is a specific website (e.g. `bbc.com`) blocked?\n* How does the blocking of different apps (e.g. WhatsApp or Telegram) vary across countries?\n* How does the blocking of a service vary across countries and [ASNs](https://ooni.org/support/glossary/#asn)?\n* How does the blocking of a service change over time?\n\nWhen trying to answer questions like the above, we normally perform relevant data analysis (instead of inspecting measurements one by one). \n\nThe MAT incorporates our data analysis techniques, enabling you to answer such questions without any data analysis skills, and with the click of a button!\n\n# How to use the MAT?\n\nThrough the filters at the start of the page, select the parameters you care about in order to plot charts based on aggregate views of OONI data.\n\nThe MAT includes the following filters:\n\n* **Countries:** Select a country through the drop-down menu (the \"All Countries\" option will show global coverage)\n* **Test Name:** Select an [OONI Probe test](https://ooni.org/nettest/) based on which you would like to get measurements (for example, select `Web Connectivity` to view the testing of websites)\n* **Domain:** Type the domain for the website you would like to get measurements (e.g. `twitter.com`)\n* **Website categories:** Select the [website category](https://github.com/citizenlab/test-lists/blob/master/lists/00-LEGEND-new_category_codes.csv) for which you would like to get measurements (e.g. `News Media` for news media websites)\n* **ASN:** Type the [ASN](https://ooni.org/support/glossary/#asn) of the network for which you would like to get measurements (e.g. `AS30722` for Vodafone Italia)\n* **Date range:** Select the date range of the measurements by adjusting the `Since` and `Until` filters\n* **Columns:** Select the values that you would like to appear on the horizontal axis of your chart\n* **Rows:** Select the values that you would like to appear on the vertical axis of your chart\n\nDepending on what you would like to explore, adjust the MAT filters accordingly and click `Show Chart`. \n\nFor example, if you would like to check the testing of BBC in all countries around the world:\n\n* Type `www.bbc.com` under `Domain`\n* Select `Countries` under the `Rows`\n* Click `Show Chart`\n\nThis will plot numerous charts based on the OONI Probe testing of `www.bbc.com` worldwide.\n\n# Interpreting MAT charts\n\nThe MAT charts (and associated tables) include the following values:\n\n* **OK count:** Successful measurements (i.e. NO sign of internet censorship)\n* **Confirmed count:** Measurements from automatically **confirmed blocked websites** (e.g. a [block page](https://ooni.org/support/glossary/#block-page) was served)\n* **Anomaly count:** Measurements that provided **signs of potential blocking** (however, [false positives](https://ooni.org/support/faq/#what-are-false-positives) can occur) \n* **Failure count:** Failed experiments that should be discarded\n* **Measurement count:** Total volume of OONI measurements (pertaining to the selected country, resource, etc.)\n\nWhen trying to identify the blocking of a service (e.g. `twitter.com`), it's useful to check whether:\n\n* Measurements are annotated as `confirmed`, automatically confirming the blocking of websites\n* A large volume of measurements (in comparison to the overall measurement count) present `anomalies` (i.e. signs of potential censorship)\n\nYou can access the raw data by clicking on the bars of charts, and subsequently clicking on the relevant measurement links. \n\n# Website categories\n\n[OONI Probe](https://ooni.org/install/) users test a wide range of [websites](https://ooni.org/support/faq/#which-websites-will-i-test-for-censorship-with-ooni-probe) that fall under the following [30 standardized categories](https://github.com/citizenlab/test-lists/blob/master/lists/00-LEGEND-new_category_codes.csv).","MAT.Help.Subtitle.Categories":"Categories","MAT.CustomTooltip.ViewMeasurements":"View measurements","ReachabilityDash.Heading.CircumventionTools":"Reachability of Censorship Circumvention Tools","ReachabilityDash.CircumventionTools.Description":"The charts below display aggregate views of OONI data based on the testing of the following circumvention tools:\n\n* [Psiphon](https://ooni.org/nettest/psiphon)\n\n* [Tor](https://ooni.org/nettest/tor)\n\n* [Tor Snowflake](https://ooni.org/nettest/tor-snowflake/)\n\nPlease note that the presence of [anomalous measurements](https://ooni.org/support/faq/#what-do-you-mean-by-anomalies) is not always indicative of blocking, as [false positives](https://ooni.org/support/faq/#what-are-false-positives) can occur. Moreover, circumvention tools often have built-in circumvention techniques for evading censorship. \n\nWe therefore recommend referring to **[Tor Metrics](https://metrics.torproject.org/)** and to the **[Psiphon Data Engine](https://psix.ca/)** to view usage stats and gain a more comprehensive understanding of whether these tools work in each country.","ReachabilityDash.Form.Label.CountrySelect.AllSelected":"All countries selected","ReachabilityDash.Form.Label.CountrySelect.SearchPlaceholder":"Search","ReachabilityDash.Form.Label.CountrySelect.SelectAll":"Select All","ReachabilityDash.Form.Label.CountrySelect.SelectAllFiltered":"Select All (Filtered)","ReachabilityDash.Form.Label.CountrySelect.InputPlaceholder":"Select Countries…","ReachabilityDash.Meta.Description":"View the accessibility of censorship circumvention tools around the world through OONI data.","DateRange.Apply":"Apply","DateRange.Cancel":"Cancel","DateRange.Today":"Today","DateRange.LastWeek":"Last Week","DateRange.LastMonth":"Last Month","DateRange.LastYear":"Last Year","ThirdPartyChart.Label.gtr":"Google (Search)","ThirdPartyChart.Label.merit-nt":"Telescope","ThirdPartyChart.Label.bgp":"BGP","ThirdPartyChart.Label.ping-slash24":"Active Probing","ThirdPartyChart.Label.cloudflare":"Cloudflare","Highlights.Political.CubaReferendum2019.Title":"2019 Constitutional Referendum","Highlights.Political.CubaReferendum2019.Text":"Blocking of independent media","Highlights.Political.VenezuelaCrisis2019.Title":"2019 Presidential Crisis","Highlights.Political.VenezuelaCrisis2019.Text":"Blocking of Wikipedia and social media","Highlights.Political.ZimbabweProtests2019.Title":"2019 Fuel Protests","Highlights.Political.ZimbabweProtests2019.Text":"Social media blocking and internet blackouts","Highlights.Political.MaliElection2018.Title":"2018 Presidential Election","Highlights.Political.MaliElection2018.Text":"Blocking of WhatsApp and Twitter","Highlights.Political.CataloniaReferendum2017.Title":"Catalonia 2017 Independence Referendum","Highlights.Political.CataloniaReferendum2017.Text":"Blocking of sites related to the referendum","Highlights.Political.IranProtests2018.Title":"2018 Anti-government Protests","Highlights.Political.IranProtests2018.Text":"Blocking of Telegram, Instagram and Tor","Highlights.Political.EthiopiaProtests2016.Title":"2016 Wave of Protests","Highlights.Political.EthiopiaProtests2016.Text":"Blocking of news websites and social media","Highlights.Political.PakistanProtests2017.Title":"2017 Protests","Highlights.Political.PakistanProtests2017.Text":"Blocking of news websites and social media","Highlights.Media.Egypt.Title":"Pervasive media censorship","Highlights.Media.Egypt.Text":"Blocking of hundreds of media websites","Highlights.Media.Venezuela.Title":"Blocking of independent media websites","Highlights.Media.Venezuela.Text":"Venezuela's economic and political crisis","Highlights.Media.SouthSudan.Title":"Blocking of foreign-based media","Highlights.Media.SouthSudan.Text":"Media accused of hostile reporting against the government","Highlights.Media.Malaysia.Title":"Blocking of media","Highlights.Media.Malaysia.Text":"1MDB scandal","Highlights.Media.Iran.Title":"Pervasive media censorship","Highlights.Media.Iran.Text":"Blocking of at least 121 news outlets","Highlights.Lgbtqi.Indonesia.Text":"Blocking of LGBTQI sites","Highlights.Lgbtqi.Iran.Text":"Blocking of Grindr","Highlights.Lgbtqi.Ethiopia.Text":"Blocking of QueerNet","Highlights.Changes.Cuba.Text":"Cuba [used to primarily serve blank block pages](https://ooni.torproject.org/post/cuba-internet-censorship-2017/), only blocking the HTTP version of websites. Now they censor access to sites that support HTTPS by means of [IP blocking](https://ooni.org/post/cuba-referendum/).","Highlights.Changes.Venezuela.Text":"Venezuelan ISPs used to primarily block sites by means of [DNS tampering](https://ooni.torproject.org/post/venezuela-internet-censorship/). Now state-owned CANTV also implements [SNI-based filtering](https://ooni.torproject.org/post/venezuela-blocking-wikipedia-and-social-media-2019/).","Highlights.Changes.Ethiopia.Text":"Ethiopia [used to block](https://ooni.org/post/ethiopia-report/) numerous news websites, LGBTQI, political opposition, and circumvention tool sites. As part of the 2018 political reforms, most of these sites have been [unblocked](https://ooni.org/post/ethiopia-unblocking/)."}}
